\name{diabetesData.5061}
\alias{diabetesData.5061}
\docType{data}
\title{
diabetesData.5061
}
\description{
E-MTAB-5061 - Single-cell RNA-seq analysis of human pancreas from type 2 diabetes patients
}
\usage{data("diabetesData.5061")}
\format{
  A data frame with 26270 variables on the following 952 observations.
  \describe{
    \item{\code{gamma.cell}}{a numeric vector}
    \item{\code{gamma.cell.1}}{a numeric vector}
    \item{\code{gamma.cell.2}}{a numeric vector}
    \item{\code{delta.cell}}{a numeric vector}
    \item{\code{alpha.cell}}{a numeric vector}
    \item{\code{alpha.cell.1}}{a numeric vector}
    \item{\code{ductal.cell}}{a numeric vector}
    \item{\code{alpha.cell.2}}{a numeric vector}
    \item{\code{alpha.cell.3}}{a numeric vector}
    \item{\code{gamma.cell.3}}{a numeric vector}
    \item{\code{alpha.cell.4}}{a numeric vector}
    \item{\code{acinar.cell}}{a numeric vector}
    \item{\code{acinar.cell.1}}{a numeric vector}
    \item{\code{alpha.cell.5}}{a numeric vector}
    \item{\code{delta.cell.1}}{a numeric vector}
    \item{\code{alpha.cell.6}}{a numeric vector}
    \item{\code{delta.cell.2}}{a numeric vector}
    \item{\code{beta.cell}}{a numeric vector}
    \item{\code{beta.cell.1}}{a numeric vector}
    \item{\code{gamma.cell.4}}{a numeric vector}
    \item{\code{ductal.cell.1}}{a numeric vector}
    \item{\code{alpha.cell.7}}{a numeric vector}
    \item{\code{gamma.cell.5}}{a numeric vector}
    \item{\code{alpha.cell.8}}{a numeric vector}
    \item{\code{alpha.cell.9}}{a numeric vector}
    \item{\code{acinar.cell.2}}{a numeric vector}
    \item{\code{gamma.cell.6}}{a numeric vector}
    \item{\code{alpha.cell.10}}{a numeric vector}
    \item{\code{beta.cell.2}}{a numeric vector}
    \item{\code{epsilon.cell}}{a numeric vector}
    \item{\code{gamma.cell.7}}{a numeric vector}
    \item{\code{gamma.cell.8}}{a numeric vector}
    \item{\code{alpha.cell.11}}{a numeric vector}
    \item{\code{alpha.cell.12}}{a numeric vector}
    \item{\code{alpha.cell.13}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell}}{a numeric vector}
    \item{\code{mast.cell}}{a numeric vector}
    \item{\code{ductal.cell.2}}{a numeric vector}
    \item{\code{alpha.cell.14}}{a numeric vector}
    \item{\code{alpha.cell.15}}{a numeric vector}
    \item{\code{beta.cell.3}}{a numeric vector}
    \item{\code{alpha.cell.16}}{a numeric vector}
    \item{\code{gamma.cell.9}}{a numeric vector}
    \item{\code{alpha.cell.17}}{a numeric vector}
    \item{\code{beta.cell.4}}{a numeric vector}
    \item{\code{ductal.cell.3}}{a numeric vector}
    \item{\code{beta.cell.5}}{a numeric vector}
    \item{\code{alpha.cell.18}}{a numeric vector}
    \item{\code{alpha.cell.19}}{a numeric vector}
    \item{\code{alpha.cell.20}}{a numeric vector}
    \item{\code{beta.cell.6}}{a numeric vector}
    \item{\code{gamma.cell.10}}{a numeric vector}
    \item{\code{alpha.cell.21}}{a numeric vector}
    \item{\code{alpha.cell.22}}{a numeric vector}
    \item{\code{beta.cell.7}}{a numeric vector}
    \item{\code{alpha.cell.23}}{a numeric vector}
    \item{\code{beta.cell.8}}{a numeric vector}
    \item{\code{co.expression.cell}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.1}}{a numeric vector}
    \item{\code{alpha.cell.24}}{a numeric vector}
    \item{\code{alpha.cell.25}}{a numeric vector}
    \item{\code{delta.cell.3}}{a numeric vector}
    \item{\code{gamma.cell.11}}{a numeric vector}
    \item{\code{gamma.cell.12}}{a numeric vector}
    \item{\code{alpha.cell.26}}{a numeric vector}
    \item{\code{alpha.cell.27}}{a numeric vector}
    \item{\code{delta.cell.4}}{a numeric vector}
    \item{\code{acinar.cell.3}}{a numeric vector}
    \item{\code{alpha.cell.28}}{a numeric vector}
    \item{\code{alpha.cell.29}}{a numeric vector}
    \item{\code{alpha.cell.30}}{a numeric vector}
    \item{\code{gamma.cell.13}}{a numeric vector}
    \item{\code{delta.cell.5}}{a numeric vector}
    \item{\code{gamma.cell.14}}{a numeric vector}
    \item{\code{alpha.cell.31}}{a numeric vector}
    \item{\code{alpha.cell.32}}{a numeric vector}
    \item{\code{alpha.cell.33}}{a numeric vector}
    \item{\code{alpha.cell.34}}{a numeric vector}
    \item{\code{delta.cell.6}}{a numeric vector}
    \item{\code{mast.cell.1}}{a numeric vector}
    \item{\code{alpha.cell.35}}{a numeric vector}
    \item{\code{beta.cell.9}}{a numeric vector}
    \item{\code{alpha.cell.36}}{a numeric vector}
    \item{\code{alpha.cell.37}}{a numeric vector}
    \item{\code{alpha.cell.38}}{a numeric vector}
    \item{\code{beta.cell.10}}{a numeric vector}
    \item{\code{alpha.cell.39}}{a numeric vector}
    \item{\code{beta.cell.11}}{a numeric vector}
    \item{\code{co.expression.cell.1}}{a numeric vector}
    \item{\code{alpha.cell.40}}{a numeric vector}
    \item{\code{alpha.cell.41}}{a numeric vector}
    \item{\code{alpha.cell.42}}{a numeric vector}
    \item{\code{gamma.cell.15}}{a numeric vector}
    \item{\code{acinar.cell.4}}{a numeric vector}
    \item{\code{delta.cell.7}}{a numeric vector}
    \item{\code{alpha.cell.43}}{a numeric vector}
    \item{\code{gamma.cell.16}}{a numeric vector}
    \item{\code{alpha.cell.44}}{a numeric vector}
    \item{\code{alpha.cell.45}}{a numeric vector}
    \item{\code{alpha.cell.46}}{a numeric vector}
    \item{\code{alpha.cell.47}}{a numeric vector}
    \item{\code{alpha.cell.48}}{a numeric vector}
    \item{\code{alpha.cell.49}}{a numeric vector}
    \item{\code{alpha.cell.50}}{a numeric vector}
    \item{\code{acinar.cell.5}}{a numeric vector}
    \item{\code{gamma.cell.17}}{a numeric vector}
    \item{\code{alpha.cell.51}}{a numeric vector}
    \item{\code{alpha.cell.52}}{a numeric vector}
    \item{\code{alpha.cell.53}}{a numeric vector}
    \item{\code{acinar.cell.6}}{a numeric vector}
    \item{\code{endothelial.cell}}{a numeric vector}
    \item{\code{delta.cell.8}}{a numeric vector}
    \item{\code{alpha.cell.54}}{a numeric vector}
    \item{\code{alpha.cell.55}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.2}}{a numeric vector}
    \item{\code{delta.cell.9}}{a numeric vector}
    \item{\code{alpha.cell.56}}{a numeric vector}
    \item{\code{alpha.cell.57}}{a numeric vector}
    \item{\code{ductal.cell.4}}{a numeric vector}
    \item{\code{gamma.cell.18}}{a numeric vector}
    \item{\code{gamma.cell.19}}{a numeric vector}
    \item{\code{alpha.cell.58}}{a numeric vector}
    \item{\code{alpha.cell.59}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.3}}{a numeric vector}
    \item{\code{alpha.cell.60}}{a numeric vector}
    \item{\code{alpha.cell.61}}{a numeric vector}
    \item{\code{gamma.cell.20}}{a numeric vector}
    \item{\code{alpha.cell.62}}{a numeric vector}
    \item{\code{ductal.cell.5}}{a numeric vector}
    \item{\code{acinar.cell.7}}{a numeric vector}
    \item{\code{alpha.cell.63}}{a numeric vector}
    \item{\code{alpha.cell.64}}{a numeric vector}
    \item{\code{alpha.cell.65}}{a numeric vector}
    \item{\code{alpha.cell.66}}{a numeric vector}
    \item{\code{alpha.cell.67}}{a numeric vector}
    \item{\code{alpha.cell.68}}{a numeric vector}
    \item{\code{gamma.cell.21}}{a numeric vector}
    \item{\code{alpha.cell.69}}{a numeric vector}
    \item{\code{acinar.cell.8}}{a numeric vector}
    \item{\code{beta.cell.12}}{a numeric vector}
    \item{\code{gamma.cell.22}}{a numeric vector}
    \item{\code{ductal.cell.6}}{a numeric vector}
    \item{\code{gamma.cell.23}}{a numeric vector}
    \item{\code{beta.cell.13}}{a numeric vector}
    \item{\code{beta.cell.14}}{a numeric vector}
    \item{\code{alpha.cell.70}}{a numeric vector}
    \item{\code{acinar.cell.9}}{a numeric vector}
    \item{\code{gamma.cell.24}}{a numeric vector}
    \item{\code{beta.cell.15}}{a numeric vector}
    \item{\code{gamma.cell.25}}{a numeric vector}
    \item{\code{co.expression.cell.2}}{a numeric vector}
    \item{\code{beta.cell.16}}{a numeric vector}
    \item{\code{beta.cell.17}}{a numeric vector}
    \item{\code{acinar.cell.10}}{a numeric vector}
    \item{\code{alpha.cell.71}}{a numeric vector}
    \item{\code{alpha.cell.72}}{a numeric vector}
    \item{\code{beta.cell.18}}{a numeric vector}
    \item{\code{alpha.cell.73}}{a numeric vector}
    \item{\code{gamma.cell.26}}{a numeric vector}
    \item{\code{beta.cell.19}}{a numeric vector}
    \item{\code{alpha.cell.74}}{a numeric vector}
    \item{\code{gamma.cell.27}}{a numeric vector}
    \item{\code{alpha.cell.75}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.4}}{a numeric vector}
    \item{\code{gamma.cell.28}}{a numeric vector}
    \item{\code{alpha.cell.76}}{a numeric vector}
    \item{\code{gamma.cell.29}}{a numeric vector}
    \item{\code{alpha.cell.77}}{a numeric vector}
    \item{\code{alpha.cell.78}}{a numeric vector}
    \item{\code{PSC.cell}}{a numeric vector}
    \item{\code{delta.cell.10}}{a numeric vector}
    \item{\code{alpha.cell.79}}{a numeric vector}
    \item{\code{alpha.cell.80}}{a numeric vector}
    \item{\code{gamma.cell.30}}{a numeric vector}
    \item{\code{gamma.cell.31}}{a numeric vector}
    \item{\code{alpha.cell.81}}{a numeric vector}
    \item{\code{alpha.cell.82}}{a numeric vector}
    \item{\code{gamma.cell.32}}{a numeric vector}
    \item{\code{ductal.cell.7}}{a numeric vector}
    \item{\code{alpha.cell.83}}{a numeric vector}
    \item{\code{alpha.cell.84}}{a numeric vector}
    \item{\code{delta.cell.11}}{a numeric vector}
    \item{\code{gamma.cell.33}}{a numeric vector}
    \item{\code{alpha.cell.85}}{a numeric vector}
    \item{\code{alpha.cell.86}}{a numeric vector}
    \item{\code{delta.cell.12}}{a numeric vector}
    \item{\code{acinar.cell.11}}{a numeric vector}
    \item{\code{alpha.cell.87}}{a numeric vector}
    \item{\code{alpha.cell.88}}{a numeric vector}
    \item{\code{alpha.cell.89}}{a numeric vector}
    \item{\code{gamma.cell.34}}{a numeric vector}
    \item{\code{beta.cell.20}}{a numeric vector}
    \item{\code{ductal.cell.8}}{a numeric vector}
    \item{\code{gamma.cell.35}}{a numeric vector}
    \item{\code{gamma.cell.36}}{a numeric vector}
    \item{\code{gamma.cell.37}}{a numeric vector}
    \item{\code{alpha.cell.90}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.5}}{a numeric vector}
    \item{\code{alpha.cell.91}}{a numeric vector}
    \item{\code{delta.cell.13}}{a numeric vector}
    \item{\code{delta.cell.14}}{a numeric vector}
    \item{\code{beta.cell.21}}{a numeric vector}
    \item{\code{ductal.cell.9}}{a numeric vector}
    \item{\code{acinar.cell.12}}{a numeric vector}
    \item{\code{delta.cell.15}}{a numeric vector}
    \item{\code{delta.cell.16}}{a numeric vector}
    \item{\code{alpha.cell.92}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.6}}{a numeric vector}
    \item{\code{beta.cell.22}}{a numeric vector}
    \item{\code{alpha.cell.93}}{a numeric vector}
    \item{\code{alpha.cell.94}}{a numeric vector}
    \item{\code{co.expression.cell.3}}{a numeric vector}
    \item{\code{alpha.cell.95}}{a numeric vector}
    \item{\code{alpha.cell.96}}{a numeric vector}
    \item{\code{alpha.cell.97}}{a numeric vector}
    \item{\code{acinar.cell.13}}{a numeric vector}
    \item{\code{delta.cell.17}}{a numeric vector}
    \item{\code{ductal.cell.10}}{a numeric vector}
    \item{\code{alpha.cell.98}}{a numeric vector}
    \item{\code{alpha.cell.99}}{a numeric vector}
    \item{\code{endothelial.cell.1}}{a numeric vector}
    \item{\code{alpha.cell.100}}{a numeric vector}
    \item{\code{alpha.cell.101}}{a numeric vector}
    \item{\code{co.expression.cell.4}}{a numeric vector}
    \item{\code{gamma.cell.38}}{a numeric vector}
    \item{\code{alpha.cell.102}}{a numeric vector}
    \item{\code{alpha.cell.103}}{a numeric vector}
    \item{\code{beta.cell.23}}{a numeric vector}
    \item{\code{delta.cell.18}}{a numeric vector}
    \item{\code{alpha.cell.104}}{a numeric vector}
    \item{\code{beta.cell.24}}{a numeric vector}
    \item{\code{alpha.cell.105}}{a numeric vector}
    \item{\code{delta.cell.19}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.7}}{a numeric vector}
    \item{\code{gamma.cell.39}}{a numeric vector}
    \item{\code{beta.cell.25}}{a numeric vector}
    \item{\code{alpha.cell.106}}{a numeric vector}
    \item{\code{acinar.cell.14}}{a numeric vector}
    \item{\code{acinar.cell.15}}{a numeric vector}
    \item{\code{PSC.cell.1}}{a numeric vector}
    \item{\code{gamma.cell.40}}{a numeric vector}
    \item{\code{MHC.class.II.cell}}{a numeric vector}
    \item{\code{beta.cell.26}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.8}}{a numeric vector}
    \item{\code{beta.cell.27}}{a numeric vector}
    \item{\code{beta.cell.28}}{a numeric vector}
    \item{\code{ductal.cell.11}}{a numeric vector}
    \item{\code{alpha.cell.107}}{a numeric vector}
    \item{\code{gamma.cell.41}}{a numeric vector}
    \item{\code{alpha.cell.108}}{a numeric vector}
    \item{\code{gamma.cell.42}}{a numeric vector}
    \item{\code{alpha.cell.109}}{a numeric vector}
    \item{\code{co.expression.cell.5}}{a numeric vector}
    \item{\code{gamma.cell.43}}{a numeric vector}
    \item{\code{beta.cell.29}}{a numeric vector}
    \item{\code{alpha.cell.110}}{a numeric vector}
    \item{\code{alpha.cell.111}}{a numeric vector}
    \item{\code{alpha.cell.112}}{a numeric vector}
    \item{\code{beta.cell.30}}{a numeric vector}
    \item{\code{alpha.cell.113}}{a numeric vector}
    \item{\code{beta.cell.31}}{a numeric vector}
    \item{\code{alpha.cell.114}}{a numeric vector}
    \item{\code{ductal.cell.12}}{a numeric vector}
    \item{\code{alpha.cell.115}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.9}}{a numeric vector}
    \item{\code{alpha.cell.116}}{a numeric vector}
    \item{\code{alpha.cell.117}}{a numeric vector}
    \item{\code{beta.cell.32}}{a numeric vector}
    \item{\code{ductal.cell.13}}{a numeric vector}
    \item{\code{alpha.cell.118}}{a numeric vector}
    \item{\code{delta.cell.20}}{a numeric vector}
    \item{\code{co.expression.cell.6}}{a numeric vector}
    \item{\code{alpha.cell.119}}{a numeric vector}
    \item{\code{ductal.cell.14}}{a numeric vector}
    \item{\code{endothelial.cell.2}}{a numeric vector}
    \item{\code{ductal.cell.15}}{a numeric vector}
    \item{\code{delta.cell.21}}{a numeric vector}
    \item{\code{alpha.cell.120}}{a numeric vector}
    \item{\code{beta.cell.33}}{a numeric vector}
    \item{\code{alpha.cell.121}}{a numeric vector}
    \item{\code{alpha.cell.122}}{a numeric vector}
    \item{\code{ductal.cell.16}}{a numeric vector}
    \item{\code{alpha.cell.123}}{a numeric vector}
    \item{\code{alpha.cell.124}}{a numeric vector}
    \item{\code{alpha.cell.125}}{a numeric vector}
    \item{\code{acinar.cell.16}}{a numeric vector}
    \item{\code{ductal.cell.17}}{a numeric vector}
    \item{\code{ductal.cell.18}}{a numeric vector}
    \item{\code{ductal.cell.19}}{a numeric vector}
    \item{\code{delta.cell.22}}{a numeric vector}
    \item{\code{alpha.cell.126}}{a numeric vector}
    \item{\code{alpha.cell.127}}{a numeric vector}
    \item{\code{alpha.cell.128}}{a numeric vector}
    \item{\code{alpha.cell.129}}{a numeric vector}
    \item{\code{delta.cell.23}}{a numeric vector}
    \item{\code{alpha.cell.130}}{a numeric vector}
    \item{\code{ductal.cell.20}}{a numeric vector}
    \item{\code{beta.cell.34}}{a numeric vector}
    \item{\code{alpha.cell.131}}{a numeric vector}
    \item{\code{alpha.cell.132}}{a numeric vector}
    \item{\code{acinar.cell.17}}{a numeric vector}
    \item{\code{alpha.cell.133}}{a numeric vector}
    \item{\code{alpha.cell.134}}{a numeric vector}
    \item{\code{delta.cell.24}}{a numeric vector}
    \item{\code{alpha.cell.135}}{a numeric vector}
    \item{\code{alpha.cell.136}}{a numeric vector}
    \item{\code{PSC.cell.2}}{a numeric vector}
    \item{\code{alpha.cell.137}}{a numeric vector}
    \item{\code{alpha.cell.138}}{a numeric vector}
    \item{\code{ductal.cell.21}}{a numeric vector}
    \item{\code{ductal.cell.22}}{a numeric vector}
    \item{\code{ductal.cell.23}}{a numeric vector}
    \item{\code{alpha.cell.139}}{a numeric vector}
    \item{\code{delta.cell.25}}{a numeric vector}
    \item{\code{gamma.cell.44}}{a numeric vector}
    \item{\code{beta.cell.35}}{a numeric vector}
    \item{\code{alpha.cell.140}}{a numeric vector}
    \item{\code{alpha.cell.141}}{a numeric vector}
    \item{\code{alpha.cell.142}}{a numeric vector}
    \item{\code{alpha.cell.143}}{a numeric vector}
    \item{\code{alpha.cell.144}}{a numeric vector}
    \item{\code{beta.cell.36}}{a numeric vector}
    \item{\code{alpha.cell.145}}{a numeric vector}
    \item{\code{alpha.cell.146}}{a numeric vector}
    \item{\code{alpha.cell.147}}{a numeric vector}
    \item{\code{PSC.cell.3}}{a numeric vector}
    \item{\code{alpha.cell.148}}{a numeric vector}
    \item{\code{alpha.cell.149}}{a numeric vector}
    \item{\code{alpha.cell.150}}{a numeric vector}
    \item{\code{delta.cell.26}}{a numeric vector}
    \item{\code{alpha.cell.151}}{a numeric vector}
    \item{\code{alpha.cell.152}}{a numeric vector}
    \item{\code{ductal.cell.24}}{a numeric vector}
    \item{\code{co.expression.cell.7}}{a numeric vector}
    \item{\code{ductal.cell.25}}{a numeric vector}
    \item{\code{ductal.cell.26}}{a numeric vector}
    \item{\code{ductal.cell.27}}{a numeric vector}
    \item{\code{alpha.cell.153}}{a numeric vector}
    \item{\code{ductal.cell.28}}{a numeric vector}
    \item{\code{alpha.cell.154}}{a numeric vector}
    \item{\code{beta.cell.37}}{a numeric vector}
    \item{\code{co.expression.cell.8}}{a numeric vector}
    \item{\code{beta.cell.38}}{a numeric vector}
    \item{\code{alpha.cell.155}}{a numeric vector}
    \item{\code{beta.cell.39}}{a numeric vector}
    \item{\code{ductal.cell.29}}{a numeric vector}
    \item{\code{ductal.cell.30}}{a numeric vector}
    \item{\code{beta.cell.40}}{a numeric vector}
    \item{\code{PSC.cell.4}}{a numeric vector}
    \item{\code{alpha.cell.156}}{a numeric vector}
    \item{\code{ductal.cell.31}}{a numeric vector}
    \item{\code{beta.cell.41}}{a numeric vector}
    \item{\code{ductal.cell.32}}{a numeric vector}
    \item{\code{beta.cell.42}}{a numeric vector}
    \item{\code{alpha.cell.157}}{a numeric vector}
    \item{\code{ductal.cell.33}}{a numeric vector}
    \item{\code{alpha.cell.158}}{a numeric vector}
    \item{\code{delta.cell.27}}{a numeric vector}
    \item{\code{alpha.cell.159}}{a numeric vector}
    \item{\code{ductal.cell.34}}{a numeric vector}
    \item{\code{endothelial.cell.3}}{a numeric vector}
    \item{\code{ductal.cell.35}}{a numeric vector}
    \item{\code{delta.cell.28}}{a numeric vector}
    \item{\code{alpha.cell.160}}{a numeric vector}
    \item{\code{alpha.cell.161}}{a numeric vector}
    \item{\code{delta.cell.29}}{a numeric vector}
    \item{\code{alpha.cell.162}}{a numeric vector}
    \item{\code{delta.cell.30}}{a numeric vector}
    \item{\code{gamma.cell.45}}{a numeric vector}
    \item{\code{alpha.cell.163}}{a numeric vector}
    \item{\code{alpha.cell.164}}{a numeric vector}
    \item{\code{gamma.cell.46}}{a numeric vector}
    \item{\code{alpha.cell.165}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.10}}{a numeric vector}
    \item{\code{PSC.cell.5}}{a numeric vector}
    \item{\code{ductal.cell.36}}{a numeric vector}
    \item{\code{ductal.cell.37}}{a numeric vector}
    \item{\code{PSC.cell.6}}{a numeric vector}
    \item{\code{ductal.cell.38}}{a numeric vector}
    \item{\code{gamma.cell.47}}{a numeric vector}
    \item{\code{delta.cell.31}}{a numeric vector}
    \item{\code{alpha.cell.166}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.11}}{a numeric vector}
    \item{\code{alpha.cell.167}}{a numeric vector}
    \item{\code{ductal.cell.39}}{a numeric vector}
    \item{\code{alpha.cell.168}}{a numeric vector}
    \item{\code{alpha.cell.169}}{a numeric vector}
    \item{\code{acinar.cell.18}}{a numeric vector}
    \item{\code{alpha.cell.170}}{a numeric vector}
    \item{\code{ductal.cell.40}}{a numeric vector}
    \item{\code{delta.cell.32}}{a numeric vector}
    \item{\code{beta.cell.43}}{a numeric vector}
    \item{\code{beta.cell.44}}{a numeric vector}
    \item{\code{ductal.cell.41}}{a numeric vector}
    \item{\code{alpha.cell.171}}{a numeric vector}
    \item{\code{delta.cell.33}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.12}}{a numeric vector}
    \item{\code{ductal.cell.42}}{a numeric vector}
    \item{\code{alpha.cell.172}}{a numeric vector}
    \item{\code{alpha.cell.173}}{a numeric vector}
    \item{\code{alpha.cell.174}}{a numeric vector}
    \item{\code{beta.cell.45}}{a numeric vector}
    \item{\code{alpha.cell.175}}{a numeric vector}
    \item{\code{alpha.cell.176}}{a numeric vector}
    \item{\code{ductal.cell.43}}{a numeric vector}
    \item{\code{ductal.cell.44}}{a numeric vector}
    \item{\code{beta.cell.46}}{a numeric vector}
    \item{\code{ductal.cell.45}}{a numeric vector}
    \item{\code{alpha.cell.177}}{a numeric vector}
    \item{\code{alpha.cell.178}}{a numeric vector}
    \item{\code{ductal.cell.46}}{a numeric vector}
    \item{\code{ductal.cell.47}}{a numeric vector}
    \item{\code{alpha.cell.179}}{a numeric vector}
    \item{\code{ductal.cell.48}}{a numeric vector}
    \item{\code{ductal.cell.49}}{a numeric vector}
    \item{\code{gamma.cell.48}}{a numeric vector}
    \item{\code{gamma.cell.49}}{a numeric vector}
    \item{\code{alpha.cell.180}}{a numeric vector}
    \item{\code{alpha.cell.181}}{a numeric vector}
    \item{\code{gamma.cell.50}}{a numeric vector}
    \item{\code{gamma.cell.51}}{a numeric vector}
    \item{\code{alpha.cell.182}}{a numeric vector}
    \item{\code{beta.cell.47}}{a numeric vector}
    \item{\code{beta.cell.48}}{a numeric vector}
    \item{\code{alpha.cell.183}}{a numeric vector}
    \item{\code{gamma.cell.52}}{a numeric vector}
    \item{\code{alpha.cell.184}}{a numeric vector}
    \item{\code{acinar.cell.19}}{a numeric vector}
    \item{\code{acinar.cell.20}}{a numeric vector}
    \item{\code{ductal.cell.50}}{a numeric vector}
    \item{\code{alpha.cell.185}}{a numeric vector}
    \item{\code{ductal.cell.51}}{a numeric vector}
    \item{\code{ductal.cell.52}}{a numeric vector}
    \item{\code{beta.cell.49}}{a numeric vector}
    \item{\code{alpha.cell.186}}{a numeric vector}
    \item{\code{alpha.cell.187}}{a numeric vector}
    \item{\code{beta.cell.50}}{a numeric vector}
    \item{\code{alpha.cell.188}}{a numeric vector}
    \item{\code{delta.cell.34}}{a numeric vector}
    \item{\code{alpha.cell.189}}{a numeric vector}
    \item{\code{ductal.cell.53}}{a numeric vector}
    \item{\code{ductal.cell.54}}{a numeric vector}
    \item{\code{alpha.cell.190}}{a numeric vector}
    \item{\code{beta.cell.51}}{a numeric vector}
    \item{\code{alpha.cell.191}}{a numeric vector}
    \item{\code{alpha.cell.192}}{a numeric vector}
    \item{\code{alpha.cell.193}}{a numeric vector}
    \item{\code{alpha.cell.194}}{a numeric vector}
    \item{\code{beta.cell.52}}{a numeric vector}
    \item{\code{gamma.cell.53}}{a numeric vector}
    \item{\code{beta.cell.53}}{a numeric vector}
    \item{\code{beta.cell.54}}{a numeric vector}
    \item{\code{alpha.cell.195}}{a numeric vector}
    \item{\code{delta.cell.35}}{a numeric vector}
    \item{\code{delta.cell.36}}{a numeric vector}
    \item{\code{ductal.cell.55}}{a numeric vector}
    \item{\code{beta.cell.55}}{a numeric vector}
    \item{\code{alpha.cell.196}}{a numeric vector}
    \item{\code{alpha.cell.197}}{a numeric vector}
    \item{\code{beta.cell.56}}{a numeric vector}
    \item{\code{ductal.cell.56}}{a numeric vector}
    \item{\code{beta.cell.57}}{a numeric vector}
    \item{\code{mast.cell.2}}{a numeric vector}
    \item{\code{beta.cell.58}}{a numeric vector}
    \item{\code{acinar.cell.21}}{a numeric vector}
    \item{\code{ductal.cell.57}}{a numeric vector}
    \item{\code{PSC.cell.7}}{a numeric vector}
    \item{\code{alpha.cell.198}}{a numeric vector}
    \item{\code{beta.cell.59}}{a numeric vector}
    \item{\code{ductal.cell.58}}{a numeric vector}
    \item{\code{beta.cell.60}}{a numeric vector}
    \item{\code{ductal.cell.59}}{a numeric vector}
    \item{\code{beta.cell.61}}{a numeric vector}
    \item{\code{gamma.cell.54}}{a numeric vector}
    \item{\code{ductal.cell.60}}{a numeric vector}
    \item{\code{alpha.cell.199}}{a numeric vector}
    \item{\code{ductal.cell.61}}{a numeric vector}
    \item{\code{alpha.cell.200}}{a numeric vector}
    \item{\code{ductal.cell.62}}{a numeric vector}
    \item{\code{delta.cell.37}}{a numeric vector}
    \item{\code{alpha.cell.201}}{a numeric vector}
    \item{\code{ductal.cell.63}}{a numeric vector}
    \item{\code{ductal.cell.64}}{a numeric vector}
    \item{\code{alpha.cell.202}}{a numeric vector}
    \item{\code{ductal.cell.65}}{a numeric vector}
    \item{\code{alpha.cell.203}}{a numeric vector}
    \item{\code{ductal.cell.66}}{a numeric vector}
    \item{\code{alpha.cell.204}}{a numeric vector}
    \item{\code{PSC.cell.8}}{a numeric vector}
    \item{\code{delta.cell.38}}{a numeric vector}
    \item{\code{delta.cell.39}}{a numeric vector}
    \item{\code{alpha.cell.205}}{a numeric vector}
    \item{\code{beta.cell.62}}{a numeric vector}
    \item{\code{ductal.cell.67}}{a numeric vector}
    \item{\code{gamma.cell.55}}{a numeric vector}
    \item{\code{beta.cell.63}}{a numeric vector}
    \item{\code{alpha.cell.206}}{a numeric vector}
    \item{\code{ductal.cell.68}}{a numeric vector}
    \item{\code{beta.cell.64}}{a numeric vector}
    \item{\code{ductal.cell.69}}{a numeric vector}
    \item{\code{ductal.cell.70}}{a numeric vector}
    \item{\code{delta.cell.40}}{a numeric vector}
    \item{\code{beta.cell.65}}{a numeric vector}
    \item{\code{alpha.cell.207}}{a numeric vector}
    \item{\code{alpha.cell.208}}{a numeric vector}
    \item{\code{acinar.cell.22}}{a numeric vector}
    \item{\code{beta.cell.66}}{a numeric vector}
    \item{\code{beta.cell.67}}{a numeric vector}
    \item{\code{ductal.cell.71}}{a numeric vector}
    \item{\code{ductal.cell.72}}{a numeric vector}
    \item{\code{delta.cell.41}}{a numeric vector}
    \item{\code{alpha.cell.209}}{a numeric vector}
    \item{\code{alpha.cell.210}}{a numeric vector}
    \item{\code{beta.cell.68}}{a numeric vector}
    \item{\code{beta.cell.69}}{a numeric vector}
    \item{\code{alpha.cell.211}}{a numeric vector}
    \item{\code{alpha.cell.212}}{a numeric vector}
    \item{\code{alpha.cell.213}}{a numeric vector}
    \item{\code{ductal.cell.73}}{a numeric vector}
    \item{\code{ductal.cell.74}}{a numeric vector}
    \item{\code{beta.cell.70}}{a numeric vector}
    \item{\code{ductal.cell.75}}{a numeric vector}
    \item{\code{alpha.cell.214}}{a numeric vector}
    \item{\code{alpha.cell.215}}{a numeric vector}
    \item{\code{alpha.cell.216}}{a numeric vector}
    \item{\code{delta.cell.42}}{a numeric vector}
    \item{\code{alpha.cell.217}}{a numeric vector}
    \item{\code{alpha.cell.218}}{a numeric vector}
    \item{\code{alpha.cell.219}}{a numeric vector}
    \item{\code{acinar.cell.23}}{a numeric vector}
    \item{\code{ductal.cell.76}}{a numeric vector}
    \item{\code{gamma.cell.56}}{a numeric vector}
    \item{\code{delta.cell.43}}{a numeric vector}
    \item{\code{ductal.cell.77}}{a numeric vector}
    \item{\code{ductal.cell.78}}{a numeric vector}
    \item{\code{gamma.cell.57}}{a numeric vector}
    \item{\code{ductal.cell.79}}{a numeric vector}
    \item{\code{ductal.cell.80}}{a numeric vector}
    \item{\code{PSC.cell.9}}{a numeric vector}
    \item{\code{beta.cell.71}}{a numeric vector}
    \item{\code{alpha.cell.220}}{a numeric vector}
    \item{\code{co.expression.cell.9}}{a numeric vector}
    \item{\code{ductal.cell.81}}{a numeric vector}
    \item{\code{delta.cell.44}}{a numeric vector}
    \item{\code{beta.cell.72}}{a numeric vector}
    \item{\code{beta.cell.73}}{a numeric vector}
    \item{\code{beta.cell.74}}{a numeric vector}
    \item{\code{delta.cell.45}}{a numeric vector}
    \item{\code{ductal.cell.82}}{a numeric vector}
    \item{\code{ductal.cell.83}}{a numeric vector}
    \item{\code{beta.cell.75}}{a numeric vector}
    \item{\code{beta.cell.76}}{a numeric vector}
    \item{\code{alpha.cell.221}}{a numeric vector}
    \item{\code{ductal.cell.84}}{a numeric vector}
    \item{\code{beta.cell.77}}{a numeric vector}
    \item{\code{alpha.cell.222}}{a numeric vector}
    \item{\code{alpha.cell.223}}{a numeric vector}
    \item{\code{ductal.cell.85}}{a numeric vector}
    \item{\code{MHC.class.II.cell.1}}{a numeric vector}
    \item{\code{ductal.cell.86}}{a numeric vector}
    \item{\code{ductal.cell.87}}{a numeric vector}
    \item{\code{ductal.cell.88}}{a numeric vector}
    \item{\code{alpha.cell.224}}{a numeric vector}
    \item{\code{alpha.cell.225}}{a numeric vector}
    \item{\code{alpha.cell.226}}{a numeric vector}
    \item{\code{alpha.cell.227}}{a numeric vector}
    \item{\code{alpha.cell.228}}{a numeric vector}
    \item{\code{acinar.cell.24}}{a numeric vector}
    \item{\code{ductal.cell.89}}{a numeric vector}
    \item{\code{MHC.class.II.cell.2}}{a numeric vector}
    \item{\code{ductal.cell.90}}{a numeric vector}
    \item{\code{ductal.cell.91}}{a numeric vector}
    \item{\code{alpha.cell.229}}{a numeric vector}
    \item{\code{alpha.cell.230}}{a numeric vector}
    \item{\code{beta.cell.78}}{a numeric vector}
    \item{\code{acinar.cell.25}}{a numeric vector}
    \item{\code{PSC.cell.10}}{a numeric vector}
    \item{\code{ductal.cell.92}}{a numeric vector}
    \item{\code{ductal.cell.93}}{a numeric vector}
    \item{\code{ductal.cell.94}}{a numeric vector}
    \item{\code{beta.cell.79}}{a numeric vector}
    \item{\code{ductal.cell.95}}{a numeric vector}
    \item{\code{alpha.cell.231}}{a numeric vector}
    \item{\code{alpha.cell.232}}{a numeric vector}
    \item{\code{beta.cell.80}}{a numeric vector}
    \item{\code{ductal.cell.96}}{a numeric vector}
    \item{\code{PSC.cell.11}}{a numeric vector}
    \item{\code{PSC.cell.12}}{a numeric vector}
    \item{\code{ductal.cell.97}}{a numeric vector}
    \item{\code{alpha.cell.233}}{a numeric vector}
    \item{\code{acinar.cell.26}}{a numeric vector}
    \item{\code{alpha.cell.234}}{a numeric vector}
    \item{\code{PSC.cell.13}}{a numeric vector}
    \item{\code{alpha.cell.235}}{a numeric vector}
    \item{\code{ductal.cell.98}}{a numeric vector}
    \item{\code{alpha.cell.236}}{a numeric vector}
    \item{\code{alpha.cell.237}}{a numeric vector}
    \item{\code{delta.cell.46}}{a numeric vector}
    \item{\code{alpha.cell.238}}{a numeric vector}
    \item{\code{ductal.cell.99}}{a numeric vector}
    \item{\code{beta.cell.81}}{a numeric vector}
    \item{\code{alpha.cell.239}}{a numeric vector}
    \item{\code{acinar.cell.27}}{a numeric vector}
    \item{\code{ductal.cell.100}}{a numeric vector}
    \item{\code{ductal.cell.101}}{a numeric vector}
    \item{\code{alpha.cell.240}}{a numeric vector}
    \item{\code{ductal.cell.102}}{a numeric vector}
    \item{\code{alpha.cell.241}}{a numeric vector}
    \item{\code{ductal.cell.103}}{a numeric vector}
    \item{\code{alpha.cell.242}}{a numeric vector}
    \item{\code{gamma.cell.58}}{a numeric vector}
    \item{\code{acinar.cell.28}}{a numeric vector}
    \item{\code{PSC.cell.14}}{a numeric vector}
    \item{\code{alpha.cell.243}}{a numeric vector}
    \item{\code{delta.cell.47}}{a numeric vector}
    \item{\code{alpha.cell.244}}{a numeric vector}
    \item{\code{alpha.cell.245}}{a numeric vector}
    \item{\code{alpha.cell.246}}{a numeric vector}
    \item{\code{alpha.cell.247}}{a numeric vector}
    \item{\code{gamma.cell.59}}{a numeric vector}
    \item{\code{alpha.cell.248}}{a numeric vector}
    \item{\code{alpha.cell.249}}{a numeric vector}
    \item{\code{ductal.cell.104}}{a numeric vector}
    \item{\code{gamma.cell.60}}{a numeric vector}
    \item{\code{alpha.cell.250}}{a numeric vector}
    \item{\code{ductal.cell.105}}{a numeric vector}
    \item{\code{ductal.cell.106}}{a numeric vector}
    \item{\code{ductal.cell.107}}{a numeric vector}
    \item{\code{ductal.cell.108}}{a numeric vector}
    \item{\code{gamma.cell.61}}{a numeric vector}
    \item{\code{beta.cell.82}}{a numeric vector}
    \item{\code{alpha.cell.251}}{a numeric vector}
    \item{\code{gamma.cell.62}}{a numeric vector}
    \item{\code{acinar.cell.29}}{a numeric vector}
    \item{\code{alpha.cell.252}}{a numeric vector}
    \item{\code{beta.cell.83}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.13}}{a numeric vector}
    \item{\code{PSC.cell.15}}{a numeric vector}
    \item{\code{delta.cell.48}}{a numeric vector}
    \item{\code{beta.cell.84}}{a numeric vector}
    \item{\code{alpha.cell.253}}{a numeric vector}
    \item{\code{ductal.cell.109}}{a numeric vector}
    \item{\code{gamma.cell.63}}{a numeric vector}
    \item{\code{ductal.cell.110}}{a numeric vector}
    \item{\code{gamma.cell.64}}{a numeric vector}
    \item{\code{ductal.cell.111}}{a numeric vector}
    \item{\code{delta.cell.49}}{a numeric vector}
    \item{\code{acinar.cell.30}}{a numeric vector}
    \item{\code{ductal.cell.112}}{a numeric vector}
    \item{\code{ductal.cell.113}}{a numeric vector}
    \item{\code{ductal.cell.114}}{a numeric vector}
    \item{\code{co.expression.cell.10}}{a numeric vector}
    \item{\code{ductal.cell.115}}{a numeric vector}
    \item{\code{alpha.cell.254}}{a numeric vector}
    \item{\code{co.expression.cell.11}}{a numeric vector}
    \item{\code{alpha.cell.255}}{a numeric vector}
    \item{\code{gamma.cell.65}}{a numeric vector}
    \item{\code{gamma.cell.66}}{a numeric vector}
    \item{\code{ductal.cell.116}}{a numeric vector}
    \item{\code{alpha.cell.256}}{a numeric vector}
    \item{\code{ductal.cell.117}}{a numeric vector}
    \item{\code{beta.cell.85}}{a numeric vector}
    \item{\code{acinar.cell.31}}{a numeric vector}
    \item{\code{ductal.cell.118}}{a numeric vector}
    \item{\code{gamma.cell.67}}{a numeric vector}
    \item{\code{delta.cell.50}}{a numeric vector}
    \item{\code{ductal.cell.119}}{a numeric vector}
    \item{\code{alpha.cell.257}}{a numeric vector}
    \item{\code{PSC.cell.16}}{a numeric vector}
    \item{\code{ductal.cell.120}}{a numeric vector}
    \item{\code{ductal.cell.121}}{a numeric vector}
    \item{\code{alpha.cell.258}}{a numeric vector}
    \item{\code{alpha.cell.259}}{a numeric vector}
    \item{\code{alpha.cell.260}}{a numeric vector}
    \item{\code{alpha.cell.261}}{a numeric vector}
    \item{\code{ductal.cell.122}}{a numeric vector}
    \item{\code{PSC.cell.17}}{a numeric vector}
    \item{\code{beta.cell.86}}{a numeric vector}
    \item{\code{alpha.cell.262}}{a numeric vector}
    \item{\code{alpha.cell.263}}{a numeric vector}
    \item{\code{ductal.cell.123}}{a numeric vector}
    \item{\code{alpha.cell.264}}{a numeric vector}
    \item{\code{acinar.cell.32}}{a numeric vector}
    \item{\code{acinar.cell.33}}{a numeric vector}
    \item{\code{ductal.cell.124}}{a numeric vector}
    \item{\code{delta.cell.51}}{a numeric vector}
    \item{\code{alpha.cell.265}}{a numeric vector}
    \item{\code{ductal.cell.125}}{a numeric vector}
    \item{\code{ductal.cell.126}}{a numeric vector}
    \item{\code{ductal.cell.127}}{a numeric vector}
    \item{\code{ductal.cell.128}}{a numeric vector}
    \item{\code{delta.cell.52}}{a numeric vector}
    \item{\code{beta.cell.87}}{a numeric vector}
    \item{\code{beta.cell.88}}{a numeric vector}
    \item{\code{alpha.cell.266}}{a numeric vector}
    \item{\code{ductal.cell.129}}{a numeric vector}
    \item{\code{beta.cell.89}}{a numeric vector}
    \item{\code{alpha.cell.267}}{a numeric vector}
    \item{\code{beta.cell.90}}{a numeric vector}
    \item{\code{beta.cell.91}}{a numeric vector}
    \item{\code{ductal.cell.130}}{a numeric vector}
    \item{\code{delta.cell.53}}{a numeric vector}
    \item{\code{ductal.cell.131}}{a numeric vector}
    \item{\code{delta.cell.54}}{a numeric vector}
    \item{\code{alpha.cell.268}}{a numeric vector}
    \item{\code{acinar.cell.34}}{a numeric vector}
    \item{\code{alpha.cell.269}}{a numeric vector}
    \item{\code{ductal.cell.132}}{a numeric vector}
    \item{\code{ductal.cell.133}}{a numeric vector}
    \item{\code{gamma.cell.68}}{a numeric vector}
    \item{\code{acinar.cell.35}}{a numeric vector}
    \item{\code{acinar.cell.36}}{a numeric vector}
    \item{\code{alpha.cell.270}}{a numeric vector}
    \item{\code{alpha.cell.271}}{a numeric vector}
    \item{\code{alpha.cell.272}}{a numeric vector}
    \item{\code{ductal.cell.134}}{a numeric vector}
    \item{\code{gamma.cell.69}}{a numeric vector}
    \item{\code{alpha.cell.273}}{a numeric vector}
    \item{\code{alpha.cell.274}}{a numeric vector}
    \item{\code{beta.cell.92}}{a numeric vector}
    \item{\code{gamma.cell.70}}{a numeric vector}
    \item{\code{acinar.cell.37}}{a numeric vector}
    \item{\code{beta.cell.93}}{a numeric vector}
    \item{\code{alpha.cell.275}}{a numeric vector}
    \item{\code{ductal.cell.135}}{a numeric vector}
    \item{\code{alpha.cell.276}}{a numeric vector}
    \item{\code{alpha.cell.277}}{a numeric vector}
    \item{\code{alpha.cell.278}}{a numeric vector}
    \item{\code{alpha.cell.279}}{a numeric vector}
    \item{\code{delta.cell.55}}{a numeric vector}
    \item{\code{beta.cell.94}}{a numeric vector}
    \item{\code{delta.cell.56}}{a numeric vector}
    \item{\code{beta.cell.95}}{a numeric vector}
    \item{\code{beta.cell.96}}{a numeric vector}
    \item{\code{alpha.cell.280}}{a numeric vector}
    \item{\code{ductal.cell.136}}{a numeric vector}
    \item{\code{alpha.cell.281}}{a numeric vector}
    \item{\code{ductal.cell.137}}{a numeric vector}
    \item{\code{co.expression.cell.12}}{a numeric vector}
    \item{\code{alpha.cell.282}}{a numeric vector}
    \item{\code{alpha.cell.283}}{a numeric vector}
    \item{\code{ductal.cell.138}}{a numeric vector}
    \item{\code{alpha.cell.284}}{a numeric vector}
    \item{\code{delta.cell.57}}{a numeric vector}
    \item{\code{ductal.cell.139}}{a numeric vector}
    \item{\code{ductal.cell.140}}{a numeric vector}
    \item{\code{ductal.cell.141}}{a numeric vector}
    \item{\code{ductal.cell.142}}{a numeric vector}
    \item{\code{ductal.cell.143}}{a numeric vector}
    \item{\code{ductal.cell.144}}{a numeric vector}
    \item{\code{alpha.cell.285}}{a numeric vector}
    \item{\code{PSC.cell.18}}{a numeric vector}
    \item{\code{acinar.cell.38}}{a numeric vector}
    \item{\code{beta.cell.97}}{a numeric vector}
    \item{\code{co.expression.cell.13}}{a numeric vector}
    \item{\code{alpha.cell.286}}{a numeric vector}
    \item{\code{alpha.cell.287}}{a numeric vector}
    \item{\code{gamma.cell.71}}{a numeric vector}
    \item{\code{beta.cell.98}}{a numeric vector}
    \item{\code{acinar.cell.39}}{a numeric vector}
    \item{\code{alpha.cell.288}}{a numeric vector}
    \item{\code{ductal.cell.145}}{a numeric vector}
    \item{\code{epsilon.cell.1}}{a numeric vector}
    \item{\code{delta.cell.58}}{a numeric vector}
    \item{\code{alpha.cell.289}}{a numeric vector}
    \item{\code{ductal.cell.146}}{a numeric vector}
    \item{\code{ductal.cell.147}}{a numeric vector}
    \item{\code{alpha.cell.290}}{a numeric vector}
    \item{\code{acinar.cell.40}}{a numeric vector}
    \item{\code{gamma.cell.72}}{a numeric vector}
    \item{\code{ductal.cell.148}}{a numeric vector}
    \item{\code{alpha.cell.291}}{a numeric vector}
    \item{\code{acinar.cell.41}}{a numeric vector}
    \item{\code{beta.cell.99}}{a numeric vector}
    \item{\code{endothelial.cell.4}}{a numeric vector}
    \item{\code{ductal.cell.149}}{a numeric vector}
    \item{\code{ductal.cell.150}}{a numeric vector}
    \item{\code{alpha.cell.292}}{a numeric vector}
    \item{\code{ductal.cell.151}}{a numeric vector}
    \item{\code{alpha.cell.293}}{a numeric vector}
    \item{\code{ductal.cell.152}}{a numeric vector}
    \item{\code{ductal.cell.153}}{a numeric vector}
    \item{\code{ductal.cell.154}}{a numeric vector}
    \item{\code{alpha.cell.294}}{a numeric vector}
    \item{\code{alpha.cell.295}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.14}}{a numeric vector}
    \item{\code{ductal.cell.155}}{a numeric vector}
    \item{\code{ductal.cell.156}}{a numeric vector}
    \item{\code{beta.cell.100}}{a numeric vector}
    \item{\code{beta.cell.101}}{a numeric vector}
    \item{\code{alpha.cell.296}}{a numeric vector}
    \item{\code{alpha.cell.297}}{a numeric vector}
    \item{\code{acinar.cell.42}}{a numeric vector}
    \item{\code{ductal.cell.157}}{a numeric vector}
    \item{\code{PSC.cell.19}}{a numeric vector}
    \item{\code{alpha.cell.298}}{a numeric vector}
    \item{\code{gamma.cell.73}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.15}}{a numeric vector}
    \item{\code{ductal.cell.158}}{a numeric vector}
    \item{\code{alpha.cell.299}}{a numeric vector}
    \item{\code{ductal.cell.159}}{a numeric vector}
    \item{\code{ductal.cell.160}}{a numeric vector}
    \item{\code{ductal.cell.161}}{a numeric vector}
    \item{\code{ductal.cell.162}}{a numeric vector}
    \item{\code{beta.cell.102}}{a numeric vector}
    \item{\code{ductal.cell.163}}{a numeric vector}
    \item{\code{ductal.cell.164}}{a numeric vector}
    \item{\code{alpha.cell.300}}{a numeric vector}
    \item{\code{alpha.cell.301}}{a numeric vector}
    \item{\code{alpha.cell.302}}{a numeric vector}
    \item{\code{alpha.cell.303}}{a numeric vector}
    \item{\code{ductal.cell.165}}{a numeric vector}
    \item{\code{gamma.cell.74}}{a numeric vector}
    \item{\code{alpha.cell.304}}{a numeric vector}
    \item{\code{alpha.cell.305}}{a numeric vector}
    \item{\code{PSC.cell.20}}{a numeric vector}
    \item{\code{delta.cell.59}}{a numeric vector}
    \item{\code{alpha.cell.306}}{a numeric vector}
    \item{\code{acinar.cell.43}}{a numeric vector}
    \item{\code{delta.cell.60}}{a numeric vector}
    \item{\code{gamma.cell.75}}{a numeric vector}
    \item{\code{acinar.cell.44}}{a numeric vector}
    \item{\code{gamma.cell.76}}{a numeric vector}
    \item{\code{alpha.cell.307}}{a numeric vector}
    \item{\code{ductal.cell.166}}{a numeric vector}
    \item{\code{beta.cell.103}}{a numeric vector}
    \item{\code{ductal.cell.167}}{a numeric vector}
    \item{\code{alpha.cell.308}}{a numeric vector}
    \item{\code{acinar.cell.45}}{a numeric vector}
    \item{\code{PSC.cell.21}}{a numeric vector}
    \item{\code{alpha.cell.309}}{a numeric vector}
    \item{\code{alpha.cell.310}}{a numeric vector}
    \item{\code{acinar.cell.46}}{a numeric vector}
    \item{\code{gamma.cell.77}}{a numeric vector}
    \item{\code{ductal.cell.168}}{a numeric vector}
    \item{\code{alpha.cell.311}}{a numeric vector}
    \item{\code{ductal.cell.169}}{a numeric vector}
    \item{\code{ductal.cell.170}}{a numeric vector}
    \item{\code{alpha.cell.312}}{a numeric vector}
    \item{\code{alpha.cell.313}}{a numeric vector}
    \item{\code{alpha.cell.314}}{a numeric vector}
    \item{\code{acinar.cell.47}}{a numeric vector}
    \item{\code{alpha.cell.315}}{a numeric vector}
    \item{\code{alpha.cell.316}}{a numeric vector}
    \item{\code{ductal.cell.171}}{a numeric vector}
    \item{\code{delta.cell.61}}{a numeric vector}
    \item{\code{ductal.cell.172}}{a numeric vector}
    \item{\code{delta.cell.62}}{a numeric vector}
    \item{\code{ductal.cell.173}}{a numeric vector}
    \item{\code{ductal.cell.174}}{a numeric vector}
    \item{\code{delta.cell.63}}{a numeric vector}
    \item{\code{alpha.cell.317}}{a numeric vector}
    \item{\code{gamma.cell.78}}{a numeric vector}
    \item{\code{alpha.cell.318}}{a numeric vector}
    \item{\code{ductal.cell.175}}{a numeric vector}
    \item{\code{ductal.cell.176}}{a numeric vector}
    \item{\code{delta.cell.64}}{a numeric vector}
    \item{\code{beta.cell.104}}{a numeric vector}
    \item{\code{alpha.cell.319}}{a numeric vector}
    \item{\code{ductal.cell.177}}{a numeric vector}
    \item{\code{alpha.cell.320}}{a numeric vector}
    \item{\code{ductal.cell.178}}{a numeric vector}
    \item{\code{ductal.cell.179}}{a numeric vector}
    \item{\code{beta.cell.105}}{a numeric vector}
    \item{\code{ductal.cell.180}}{a numeric vector}
    \item{\code{ductal.cell.181}}{a numeric vector}
    \item{\code{delta.cell.65}}{a numeric vector}
    \item{\code{beta.cell.106}}{a numeric vector}
    \item{\code{gamma.cell.79}}{a numeric vector}
    \item{\code{delta.cell.66}}{a numeric vector}
    \item{\code{ductal.cell.182}}{a numeric vector}
    \item{\code{gamma.cell.80}}{a numeric vector}
    \item{\code{ductal.cell.183}}{a numeric vector}
    \item{\code{beta.cell.107}}{a numeric vector}
    \item{\code{beta.cell.108}}{a numeric vector}
    \item{\code{alpha.cell.321}}{a numeric vector}
    \item{\code{ductal.cell.184}}{a numeric vector}
    \item{\code{ductal.cell.185}}{a numeric vector}
    \item{\code{alpha.cell.322}}{a numeric vector}
    \item{\code{ductal.cell.186}}{a numeric vector}
    \item{\code{ductal.cell.187}}{a numeric vector}
    \item{\code{alpha.cell.323}}{a numeric vector}
    \item{\code{gamma.cell.81}}{a numeric vector}
    \item{\code{acinar.cell.48}}{a numeric vector}
    \item{\code{alpha.cell.324}}{a numeric vector}
    \item{\code{gamma.cell.82}}{a numeric vector}
    \item{\code{ductal.cell.188}}{a numeric vector}
    \item{\code{beta.cell.109}}{a numeric vector}
    \item{\code{beta.cell.110}}{a numeric vector}
    \item{\code{gamma.cell.83}}{a numeric vector}
    \item{\code{gamma.cell.84}}{a numeric vector}
    \item{\code{alpha.cell.325}}{a numeric vector}
    \item{\code{acinar.cell.49}}{a numeric vector}
    \item{\code{ductal.cell.189}}{a numeric vector}
    \item{\code{gamma.cell.85}}{a numeric vector}
    \item{\code{acinar.cell.50}}{a numeric vector}
    \item{\code{ductal.cell.190}}{a numeric vector}
    \item{\code{ductal.cell.191}}{a numeric vector}
    \item{\code{alpha.cell.326}}{a numeric vector}
    \item{\code{alpha.cell.327}}{a numeric vector}
    \item{\code{alpha.cell.328}}{a numeric vector}
    \item{\code{beta.cell.111}}{a numeric vector}
    \item{\code{ductal.cell.192}}{a numeric vector}
    \item{\code{ductal.cell.193}}{a numeric vector}
    \item{\code{ductal.cell.194}}{a numeric vector}
    \item{\code{ductal.cell.195}}{a numeric vector}
    \item{\code{alpha.cell.329}}{a numeric vector}
    \item{\code{alpha.cell.330}}{a numeric vector}
    \item{\code{ductal.cell.196}}{a numeric vector}
    \item{\code{ductal.cell.197}}{a numeric vector}
    \item{\code{alpha.cell.331}}{a numeric vector}
    \item{\code{alpha.cell.332}}{a numeric vector}
    \item{\code{ductal.cell.198}}{a numeric vector}
    \item{\code{co.expression.cell.14}}{a numeric vector}
    \item{\code{beta.cell.112}}{a numeric vector}
    \item{\code{alpha.cell.333}}{a numeric vector}
    \item{\code{beta.cell.113}}{a numeric vector}
    \item{\code{acinar.cell.51}}{a numeric vector}
    \item{\code{beta.cell.114}}{a numeric vector}
    \item{\code{ductal.cell.199}}{a numeric vector}
    \item{\code{acinar.cell.52}}{a numeric vector}
    \item{\code{ductal.cell.200}}{a numeric vector}
    \item{\code{alpha.cell.334}}{a numeric vector}
    \item{\code{alpha.cell.335}}{a numeric vector}
    \item{\code{ductal.cell.201}}{a numeric vector}
    \item{\code{beta.cell.115}}{a numeric vector}
    \item{\code{co.expression.cell.15}}{a numeric vector}
    \item{\code{ductal.cell.202}}{a numeric vector}
    \item{\code{alpha.cell.336}}{a numeric vector}
    \item{\code{beta.cell.116}}{a numeric vector}
    \item{\code{beta.cell.117}}{a numeric vector}
    \item{\code{ductal.cell.203}}{a numeric vector}
    \item{\code{ductal.cell.204}}{a numeric vector}
    \item{\code{gamma.cell.86}}{a numeric vector}
    \item{\code{ductal.cell.205}}{a numeric vector}
    \item{\code{alpha.cell.337}}{a numeric vector}
    \item{\code{gamma.cell.87}}{a numeric vector}
    \item{\code{alpha.cell.338}}{a numeric vector}
    \item{\code{gamma.cell.88}}{a numeric vector}
    \item{\code{delta.cell.67}}{a numeric vector}
    \item{\code{ductal.cell.206}}{a numeric vector}
    \item{\code{alpha.cell.339}}{a numeric vector}
    \item{\code{alpha.cell.340}}{a numeric vector}
    \item{\code{delta.cell.68}}{a numeric vector}
    \item{\code{acinar.cell.53}}{a numeric vector}
    \item{\code{acinar.cell.54}}{a numeric vector}
    \item{\code{alpha.cell.341}}{a numeric vector}
    \item{\code{gamma.cell.89}}{a numeric vector}
    \item{\code{alpha.cell.342}}{a numeric vector}
    \item{\code{delta.cell.69}}{a numeric vector}
    \item{\code{alpha.cell.343}}{a numeric vector}
    \item{\code{alpha.cell.344}}{a numeric vector}
  }
}
\details{
                acinar.cell                  alpha.cell                   beta.cell
                         55                         345                         118
         co.expression.cell                  delta.cell                 ductal.cell
                         16                          70                         207
           endothelial.cell                epsilon.cell                  gamma.cell
                          5                           2                          90
                  mast.cell           MHC.class.II.cell                    PSC.cell
                          3                           3                          22
unclassified.endocrine.cell
                         16
}
\source{
E-MTAB-5061 - Single-cell RNA-seq analysis of human pancreas from healthy individuals and type 2 diabetes patients

https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-5061/

Segerstolpe, Å., Palasantza, A., Eliasson, P., Andersson, E.M., Andréasson, A.C., Sun, X., Picelli, S., Sabirsh, A., Clausen, M., Bjursell, M.K. and Smith, D.M., 2016. Single-cell transcriptome profiling of human pancreatic islets in health and type 2 diabetes. Cell metabolism, 24(4), pp.593-607.
}
\examples{
cellCounts <- enumerateCellTypes(ADAPTSdata3::diabetesData.5061)
}
\keyword{datasets}
