% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADAPTSdata3.R
\name{enumerateCellTypes}
\alias{enumerateCellTypes}
\title{Build a table that shows the counts of cell types in the ADAPTSdata3 single cell data}
\usage{
enumerateCellTypes(dataFile)
}
\arguments{
\item{dataFile}{The sing cell data file}
}
\value{
a table
}
\description{
Build a table that shows the counts of cell types in the ADAPTSdata3 single cell data
}
\examples{
cellCounts <- enumerateCellTypes(ADAPTSdata3::normalData.5061)
}
