\name{gList}
\alias{gList}
\docType{data}
\title{
List of Differentially Expressed Genes for ADAPTS Vignette #2
}
\description{
The ranked list of differentially expressed genes for the normal pancreatic cells

gList <- ADAPTS::rankByT(normalData,remZinf = TRUE)

Provided here to avoid long (15+ minute) calculations
}
\usage{data("gList")}
\format{
  The format is:
List of 13
 $ gamma.cell                 :'data.frame':	9254 obs. of  4 variables:
  ..$ rat : num [1:9254] 0.987 1.015 1.017 1.019 0.98 ...
  ..$ t   : num [1:9254] -1.84 1.5 1.55 1.82 -2.04 ...
  ..$ pVal: num [1:9254] 0.0683 0.1347 0.1222 0.0698 0.0424 ...
  ..$ qVal: num [1:9254] 0.153 0.261 0.242 0.156 0.105 ...
 $ alpha.cell                 :'data.frame':	13104 obs. of  4 variables:
  ..$ rat : num [1:13104] 0.993 1.01 0.989 0.989 1.012 ...
  ..$ t   : num [1:13104] -1.39 1.37 -1.38 -1.36 1.31 ...
  ..$ pVal: num [1:13104] 0.165 0.172 0.166 0.174 0.192 ...
  ..$ qVal: num [1:13104] 0.256 0.266 0.258 0.268 0.29 ...
 $ beta.cell                  :'data.frame':	10570 obs. of  4 variables:
  ..$ rat : num [1:10570] 0.987 1.014 1.014 1.015 1.016 ...
  ..$ t   : num [1:10570] -1.45 1.7 1.9 1.4 1.36 ...
  ..$ pVal: num [1:10570] 0.1487 0.0904 0.0582 0.1642 0.1746 ...
  ..$ qVal: num [1:10570] 0.256 0.172 0.119 0.278 0.292 ...
 $ acinar.cell                :'data.frame':	13224 obs. of  4 variables:
  ..$ rat : num [1:13224] 1.011 0.989 1.012 1.013 0.985 ...
  ..$ t   : num [1:13224] 1.61 -1.4 1.52 1.47 -2.37 ...
  ..$ pVal: num [1:13224] 0.1092 0.1623 0.13 0.1436 0.0188 ...
  ..$ qVal: num [1:13224] 0.1693 0.2379 0.1968 0.2144 0.0357 ...
 $ epsilon.cell               :'data.frame':	585 obs. of  4 variables:
  ..$ rat : num [1:585] 1.09 1.1 1.11 1.13 1.14 ...
  ..$ t   : num [1:585] 4.84 3.93 4.92 4.11 4.64 ...
  ..$ pVal: num [1:585] 0.00456 0.01415 0.00519 0.0126 0.00768 ...
  ..$ qVal: num [1:585] 0.139 0.273 0.152 0.254 0.187 ...
 $ ductal.cell                :'data.frame':	13809 obs. of  4 variables:
  ..$ rat : num [1:13809] 1.01 1.011 0.988 1.012 1.014 ...
  ..$ t   : num [1:13809] 1.89 1.65 -1.64 1.44 1.34 ...
  ..$ pVal: num [1:13809] 0.0596 0.1002 0.1014 0.151 0.1805 ...
  ..$ qVal: num [1:13809] 0.0942 0.1499 0.1514 0.2148 0.2505 ...
 $ delta.cell                 :'data.frame':	6170 obs. of  4 variables:
  ..$ rat : num [1:6170] 1.017 0.98 0.978 0.978 0.976 ...
  ..$ t   : num [1:6170] 1.77 -1.78 -1.87 -1.67 -2.1 ...
  ..$ pVal: num [1:6170] 0.0824 0.0801 0.067 0.1016 0.0406 ...
  ..$ qVal: num [1:6170] 0.234 0.229 0.201 0.27 0.14 ...
 $ unclassified.endocrine.cell:'data.frame':	9438 obs. of  4 variables:
  ..$ rat : num [1:9438] 0.875 0.858 0.857 0.848 0.844 ...
  ..$ t   : num [1:9438] -1.35 -1.61 -3.46 -2.14 -4.32 ...
  ..$ pVal: num [1:9438] 0.18911 0.11956 0.00204 0.04243 0.00023 ...
  ..$ qVal: num [1:9438] 0.228534 0.148155 0.003232 0.055751 0.000426 ...
 $ endothelial.cell           :'data.frame':	4264 obs. of  4 variables:
  ..$ rat : num [1:4264] 1.04 1.04 1.05 1.05 1.05 ...
  ..$ t   : num [1:4264] 1.72 2.22 2.89 1.98 1.78 ...
  ..$ pVal: num [1:4264] 0.1116 0.0487 0.012 0.0747 0.1039 ...
  ..$ qVal: num [1:4264] 0.2942 0.1658 0.0591 0.2225 0.28 ...
 $ co.expression.cell         :'data.frame':	4434 obs. of  4 variables:
  ..$ rat : num [1:4434] 1.023 0.977 1.025 0.974 1.028 ...
  ..$ t   : num [1:4434] 1.86 -1.83 2.32 -1.94 2.42 ...
  ..$ pVal: num [1:4434] 0.0724 0.0796 0.0271 0.0612 0.0229 ...
  ..$ qVal: num [1:4434] 0.281 0.293 0.155 0.255 0.139 ...
 $ PSC.cell                   :'data.frame':	7497 obs. of  4 variables:
  ..$ rat : num [1:7497] 1.026 1.028 0.97 1.033 0.968 ...
  ..$ t   : num [1:7497] 1.49 1.97 -1.51 1.63 -1.62 ...
  ..$ pVal: num [1:7497] 0.1452 0.0573 0.1406 0.1124 0.1148 ...
  ..$ qVal: num [1:7497] 0.295 0.146 0.288 0.244 0.248 ...
 $ mast.cell                  :'data.frame':	996 obs. of  4 variables:
  ..$ rat : num [1:996] 0.969 1.047 0.95 1.075 1.076 ...
  ..$ t   : num [1:996] -3.85 7.48 -3.24 3.68 3.93 ...
  ..$ pVal: num [1:996] 0.020392 0.000201 0.037894 0.027516 0.026181 ...
  ..$ qVal: num [1:996] 0.17029 0.00822 0.26728 0.212 0.20415 ...
 $ MHC.class.II.cell          :'data.frame':	391 obs. of  4 variables:
  ..$ rat : num [1:391] 0.973 0.957 1.054 1.06 1.068 ...
  ..$ t   : num [1:391] -5.45 -3.25 6.92 3.15 9.37 ...
  ..$ pVal: num [1:391] 8.03e-08 3.68e-03 4.53e-11 3.39e-03 1.55e-08 ...
  ..$ qVal: num [1:391] 5.47e-06 1.12e-01 3.78e-09 1.05e-01 1.11e-06 ...
}
\details{
ADAPTS Vignette #2: Single Cell Analysis
}
\examples{
summary(gList)
}
\keyword{datasets}
