\name{hier.top100}
\alias{hier.top100}
\docType{data}
\title{
Top 100 Gene Hierarchical Splits of Normal Pancrease
}
\description{
The hierarchical data for the augmented deconvolution of the normal data in E-MTAB-506

The results of ADAPTS::hierarchicalSplit(sigMatrix = topAug.var100$sigMatrix, geneExpr = allSCdata) in the ADAPTS single cell vignette

Provided here to avoid long (15+ minute) calculations
}
\usage{data("hier.top100")}
\format{
  The format is:
List of 3
 $ allClusters  :List of 7
  ..$ : chr [1:2] "acinar.cell" "ductal.cell"
  ..$ : chr [1:2] "alpha.cell" "gamma.cell"
  ..$ : chr [1:4] "beta.cell" "co.expression.cell" "delta.cell" "unclassified.endocrine.cell"
  ..$ : chr [1:2] "endothelial.cell" "PSC.cell"
  ..$ : chr "epsilon.cell"
  ..$ : chr "mast.cell"
  ..$ : chr "MHC.class.II.cell"
 $ sigMatList   :List of 7
  ..$ :'data.frame':	11 obs. of  2 variables:
  .. ..$ acinar.cell: num [1:11] 9.662 0.949 11.307 8.586 8.875 ...
  .. ..$ ductal.cell: num [1:11] 4.02 6.72 5.13 2.3 2.53 ...
  ..$ :'data.frame':	11 obs. of  2 variables:
  .. ..$ alpha.cell: num [1:11] 11.26 3.31 6.62 5.06 3.95 ...
  .. ..$ gamma.cell: num [1:11] 9.38 5.48 4.41 7.38 1.21 ...
  ..$ :'data.frame':	206 obs. of  4 variables:
  .. ..$ beta.cell                  : num [1:206] 0.714 5.576 5.878 0.61 5.18 ...
  .. ..$ co.expression.cell         : num [1:206] 4.7 6.08 7.15 5.19 5.02 ...
  .. ..$ delta.cell                 : num [1:206] 5.053 5.959 8.936 0.772 2.617 ...
  .. ..$ unclassified.endocrine.cell: num [1:206] 2.947 1.908 3.881 2.556 0.365 ...
  ..$ :'data.frame':	13 obs. of  2 variables:
  .. ..$ endothelial.cell: num [1:13] 5.513 0.964 3.413 1.304 0.917 ...
  .. ..$ PSC.cell        : num [1:13] 2.3398 4.148 0.0546 5.095 4.7487 ...
  ..$ : num [1:100, 1:13] 6.244 4.4 0.163 0.598 3.626 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:100] "CD24" "PPIC" "GAD2" "RASD1" ...
  .. .. ..$ : chr [1:13] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  ..$ : num [1:100, 1:13] 6.244 4.4 0.163 0.598 3.626 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:100] "CD24" "PPIC" "GAD2" "RASD1" ...
  .. .. ..$ : chr [1:13] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  ..$ : num [1:100, 1:13] 6.244 4.4 0.163 0.598 3.626 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:100] "CD24" "PPIC" "GAD2" "RASD1" ...
  .. .. ..$ : chr [1:13] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
 $ deconMatrices:List of 20
  ..$ : num [1:100, 1:13] 6.244 4.4 0.163 0.598 3.626 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:100] "CD24" "PPIC" "GAD2" "RASD1" ...
  .. .. ..$ : chr [1:13] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  ..$ : num [1:14, 1:13] 62.74831 0.23338 0.00908 0.08915 0 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:14] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  .. .. ..$ : chr [1:13] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  ..$ : num [1:14, 1:13] 45.1688 0.0777 0 0 0 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:14] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  .. .. ..$ : chr [1:13] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  ..$ : num [1:14, 1:13] 39.7 0 0 0 0 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:14] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  .. .. ..$ : chr [1:13] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  ..$ : num [1:14, 1:13] 40.3 0 0 0 0 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:14] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  .. .. ..$ : chr [1:13] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  ..$ : num [1:14, 1:13] 38.4 0 0 0 0 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:14] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  .. .. ..$ : chr [1:13] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  ..$ : num [1:14, 1:13] 39 0 0 0 0 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:14] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  .. .. ..$ : chr [1:13] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  ..$ : num [1:14, 1:13] 42.5 0 0 0 0 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:14] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  .. .. ..$ : chr [1:13] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  ..$ : num [1:14, 1:13] 45.4 0 0 0 0 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:14] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  .. .. ..$ : chr [1:13] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  ..$ : num [1:14, 1:13] 48.3 0 0 0 0 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:14] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  .. .. ..$ : chr [1:13] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  ..$ : num [1:14, 1:13] 49.6 0 0 0 0 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:14] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  .. .. ..$ : chr [1:13] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  ..$ : num [1:14, 1:13] 49.6 0 0 0 0 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:14] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  .. .. ..$ : chr [1:13] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  ..$ : num [1:14, 1:13] 49.7 0 0 0 0 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:14] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  .. .. ..$ : chr [1:13] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  ..$ : num [1:14, 1:13] 49.7 0 0 0 0 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:14] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  .. .. ..$ : chr [1:13] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  ..$ : num [1:14, 1:13] 49.7 0 0 0 0 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:14] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  .. .. ..$ : chr [1:13] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  ..$ : num [1:14, 1:13] 49.7 0 0 0 0 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:14] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  .. .. ..$ : chr [1:13] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  ..$ : num [1:14, 1:13] 49.7 0 0 0 0 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:14] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  .. .. ..$ : chr [1:13] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  ..$ : num [1:14, 1:13] 49.7 0 0 0 0 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:14] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  .. .. ..$ : chr [1:13] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  ..$ : num [1:14, 1:13] 49.7 0 0 0 0 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:14] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  .. .. ..$ : chr [1:13] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  ..$ : num [1:14, 1:13] 49.7 0 0 0 0 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:14] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
  .. .. ..$ : chr [1:13] "acinar.cell" "alpha.cell" "beta.cell" "co.expression.cell" ...
}
\details{
ADAPTS Vignette #2: Single Cell Analysis
}
\examples{
summary(hier.top100)
}
\keyword{datasets}
