\name{normalData.5061}
\alias{normalData.5061}
\docType{data}
\title{
normalData.5061
}
\description{
E-MTAB-5061 - Single-cell RNA-seq analysis of human pancreas from healthy individuals
}
\usage{data("normalData.5061")}
\format{
  A data frame with 26270 variables on the following 1255 observations.
  \describe{
    \item{\code{gamma.cell}}{a numeric vector}
    \item{\code{alpha.cell}}{a numeric vector}
    \item{\code{beta.cell}}{a numeric vector}
    \item{\code{acinar.cell}}{a numeric vector}
    \item{\code{alpha.cell.1}}{a numeric vector}
    \item{\code{beta.cell.1}}{a numeric vector}
    \item{\code{epsilon.cell}}{a numeric vector}
    \item{\code{gamma.cell.1}}{a numeric vector}
    \item{\code{alpha.cell.2}}{a numeric vector}
    \item{\code{gamma.cell.2}}{a numeric vector}
    \item{\code{acinar.cell.1}}{a numeric vector}
    \item{\code{acinar.cell.2}}{a numeric vector}
    \item{\code{alpha.cell.3}}{a numeric vector}
    \item{\code{beta.cell.2}}{a numeric vector}
    \item{\code{alpha.cell.4}}{a numeric vector}
    \item{\code{gamma.cell.3}}{a numeric vector}
    \item{\code{gamma.cell.4}}{a numeric vector}
    \item{\code{alpha.cell.5}}{a numeric vector}
    \item{\code{alpha.cell.6}}{a numeric vector}
    \item{\code{ductal.cell}}{a numeric vector}
    \item{\code{alpha.cell.7}}{a numeric vector}
    \item{\code{alpha.cell.8}}{a numeric vector}
    \item{\code{gamma.cell.5}}{a numeric vector}
    \item{\code{acinar.cell.3}}{a numeric vector}
    \item{\code{alpha.cell.9}}{a numeric vector}
    \item{\code{alpha.cell.10}}{a numeric vector}
    \item{\code{alpha.cell.11}}{a numeric vector}
    \item{\code{beta.cell.3}}{a numeric vector}
    \item{\code{alpha.cell.12}}{a numeric vector}
    \item{\code{beta.cell.4}}{a numeric vector}
    \item{\code{alpha.cell.13}}{a numeric vector}
    \item{\code{gamma.cell.6}}{a numeric vector}
    \item{\code{alpha.cell.14}}{a numeric vector}
    \item{\code{beta.cell.5}}{a numeric vector}
    \item{\code{ductal.cell.1}}{a numeric vector}
    \item{\code{delta.cell}}{a numeric vector}
    \item{\code{alpha.cell.15}}{a numeric vector}
    \item{\code{beta.cell.6}}{a numeric vector}
    \item{\code{alpha.cell.16}}{a numeric vector}
    \item{\code{gamma.cell.7}}{a numeric vector}
    \item{\code{alpha.cell.17}}{a numeric vector}
    \item{\code{beta.cell.7}}{a numeric vector}
    \item{\code{beta.cell.8}}{a numeric vector}
    \item{\code{alpha.cell.18}}{a numeric vector}
    \item{\code{ductal.cell.2}}{a numeric vector}
    \item{\code{gamma.cell.8}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell}}{a numeric vector}
    \item{\code{alpha.cell.19}}{a numeric vector}
    \item{\code{beta.cell.9}}{a numeric vector}
    \item{\code{alpha.cell.20}}{a numeric vector}
    \item{\code{alpha.cell.21}}{a numeric vector}
    \item{\code{alpha.cell.22}}{a numeric vector}
    \item{\code{alpha.cell.23}}{a numeric vector}
    \item{\code{alpha.cell.24}}{a numeric vector}
    \item{\code{alpha.cell.25}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.1}}{a numeric vector}
    \item{\code{gamma.cell.9}}{a numeric vector}
    \item{\code{gamma.cell.10}}{a numeric vector}
    \item{\code{alpha.cell.26}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.2}}{a numeric vector}
    \item{\code{alpha.cell.27}}{a numeric vector}
    \item{\code{alpha.cell.28}}{a numeric vector}
    \item{\code{gamma.cell.11}}{a numeric vector}
    \item{\code{alpha.cell.29}}{a numeric vector}
    \item{\code{gamma.cell.12}}{a numeric vector}
    \item{\code{gamma.cell.13}}{a numeric vector}
    \item{\code{beta.cell.10}}{a numeric vector}
    \item{\code{alpha.cell.30}}{a numeric vector}
    \item{\code{alpha.cell.31}}{a numeric vector}
    \item{\code{alpha.cell.32}}{a numeric vector}
    \item{\code{alpha.cell.33}}{a numeric vector}
    \item{\code{alpha.cell.34}}{a numeric vector}
    \item{\code{alpha.cell.35}}{a numeric vector}
    \item{\code{acinar.cell.4}}{a numeric vector}
    \item{\code{alpha.cell.36}}{a numeric vector}
    \item{\code{acinar.cell.5}}{a numeric vector}
    \item{\code{alpha.cell.37}}{a numeric vector}
    \item{\code{alpha.cell.38}}{a numeric vector}
    \item{\code{alpha.cell.39}}{a numeric vector}
    \item{\code{gamma.cell.14}}{a numeric vector}
    \item{\code{alpha.cell.40}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.3}}{a numeric vector}
    \item{\code{beta.cell.11}}{a numeric vector}
    \item{\code{alpha.cell.41}}{a numeric vector}
    \item{\code{alpha.cell.42}}{a numeric vector}
    \item{\code{alpha.cell.43}}{a numeric vector}
    \item{\code{alpha.cell.44}}{a numeric vector}
    \item{\code{alpha.cell.45}}{a numeric vector}
    \item{\code{beta.cell.12}}{a numeric vector}
    \item{\code{alpha.cell.46}}{a numeric vector}
    \item{\code{ductal.cell.3}}{a numeric vector}
    \item{\code{delta.cell.1}}{a numeric vector}
    \item{\code{alpha.cell.47}}{a numeric vector}
    \item{\code{alpha.cell.48}}{a numeric vector}
    \item{\code{alpha.cell.49}}{a numeric vector}
    \item{\code{acinar.cell.6}}{a numeric vector}
    \item{\code{alpha.cell.50}}{a numeric vector}
    \item{\code{ductal.cell.4}}{a numeric vector}
    \item{\code{alpha.cell.51}}{a numeric vector}
    \item{\code{alpha.cell.52}}{a numeric vector}
    \item{\code{endothelial.cell}}{a numeric vector}
    \item{\code{alpha.cell.53}}{a numeric vector}
    \item{\code{gamma.cell.15}}{a numeric vector}
    \item{\code{beta.cell.13}}{a numeric vector}
    \item{\code{acinar.cell.7}}{a numeric vector}
    \item{\code{alpha.cell.54}}{a numeric vector}
    \item{\code{alpha.cell.55}}{a numeric vector}
    \item{\code{alpha.cell.56}}{a numeric vector}
    \item{\code{alpha.cell.57}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.4}}{a numeric vector}
    \item{\code{alpha.cell.58}}{a numeric vector}
    \item{\code{alpha.cell.59}}{a numeric vector}
    \item{\code{alpha.cell.60}}{a numeric vector}
    \item{\code{alpha.cell.61}}{a numeric vector}
    \item{\code{gamma.cell.16}}{a numeric vector}
    \item{\code{delta.cell.2}}{a numeric vector}
    \item{\code{alpha.cell.62}}{a numeric vector}
    \item{\code{alpha.cell.63}}{a numeric vector}
    \item{\code{alpha.cell.64}}{a numeric vector}
    \item{\code{alpha.cell.65}}{a numeric vector}
    \item{\code{delta.cell.3}}{a numeric vector}
    \item{\code{alpha.cell.66}}{a numeric vector}
    \item{\code{alpha.cell.67}}{a numeric vector}
    \item{\code{alpha.cell.68}}{a numeric vector}
    \item{\code{alpha.cell.69}}{a numeric vector}
    \item{\code{alpha.cell.70}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.5}}{a numeric vector}
    \item{\code{alpha.cell.71}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.6}}{a numeric vector}
    \item{\code{co.expression.cell}}{a numeric vector}
    \item{\code{alpha.cell.72}}{a numeric vector}
    \item{\code{alpha.cell.73}}{a numeric vector}
    \item{\code{beta.cell.14}}{a numeric vector}
    \item{\code{alpha.cell.74}}{a numeric vector}
    \item{\code{alpha.cell.75}}{a numeric vector}
    \item{\code{gamma.cell.17}}{a numeric vector}
    \item{\code{gamma.cell.18}}{a numeric vector}
    \item{\code{ductal.cell.5}}{a numeric vector}
    \item{\code{gamma.cell.19}}{a numeric vector}
    \item{\code{alpha.cell.76}}{a numeric vector}
    \item{\code{gamma.cell.20}}{a numeric vector}
    \item{\code{alpha.cell.77}}{a numeric vector}
    \item{\code{alpha.cell.78}}{a numeric vector}
    \item{\code{alpha.cell.79}}{a numeric vector}
    \item{\code{alpha.cell.80}}{a numeric vector}
    \item{\code{alpha.cell.81}}{a numeric vector}
    \item{\code{alpha.cell.82}}{a numeric vector}
    \item{\code{alpha.cell.83}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.7}}{a numeric vector}
    \item{\code{beta.cell.15}}{a numeric vector}
    \item{\code{gamma.cell.21}}{a numeric vector}
    \item{\code{gamma.cell.22}}{a numeric vector}
    \item{\code{gamma.cell.23}}{a numeric vector}
    \item{\code{alpha.cell.84}}{a numeric vector}
    \item{\code{alpha.cell.85}}{a numeric vector}
    \item{\code{alpha.cell.86}}{a numeric vector}
    \item{\code{alpha.cell.87}}{a numeric vector}
    \item{\code{alpha.cell.88}}{a numeric vector}
    \item{\code{delta.cell.4}}{a numeric vector}
    \item{\code{beta.cell.16}}{a numeric vector}
    \item{\code{delta.cell.5}}{a numeric vector}
    \item{\code{gamma.cell.24}}{a numeric vector}
    \item{\code{alpha.cell.89}}{a numeric vector}
    \item{\code{alpha.cell.90}}{a numeric vector}
    \item{\code{alpha.cell.91}}{a numeric vector}
    \item{\code{alpha.cell.92}}{a numeric vector}
    \item{\code{delta.cell.6}}{a numeric vector}
    \item{\code{alpha.cell.93}}{a numeric vector}
    \item{\code{beta.cell.17}}{a numeric vector}
    \item{\code{alpha.cell.94}}{a numeric vector}
    \item{\code{acinar.cell.8}}{a numeric vector}
    \item{\code{alpha.cell.95}}{a numeric vector}
    \item{\code{ductal.cell.6}}{a numeric vector}
    \item{\code{beta.cell.18}}{a numeric vector}
    \item{\code{alpha.cell.96}}{a numeric vector}
    \item{\code{delta.cell.7}}{a numeric vector}
    \item{\code{beta.cell.19}}{a numeric vector}
    \item{\code{beta.cell.20}}{a numeric vector}
    \item{\code{gamma.cell.25}}{a numeric vector}
    \item{\code{alpha.cell.97}}{a numeric vector}
    \item{\code{alpha.cell.98}}{a numeric vector}
    \item{\code{alpha.cell.99}}{a numeric vector}
    \item{\code{gamma.cell.26}}{a numeric vector}
    \item{\code{alpha.cell.100}}{a numeric vector}
    \item{\code{alpha.cell.101}}{a numeric vector}
    \item{\code{alpha.cell.102}}{a numeric vector}
    \item{\code{delta.cell.8}}{a numeric vector}
    \item{\code{acinar.cell.9}}{a numeric vector}
    \item{\code{alpha.cell.103}}{a numeric vector}
    \item{\code{alpha.cell.104}}{a numeric vector}
    \item{\code{alpha.cell.105}}{a numeric vector}
    \item{\code{gamma.cell.27}}{a numeric vector}
    \item{\code{alpha.cell.106}}{a numeric vector}
    \item{\code{alpha.cell.107}}{a numeric vector}
    \item{\code{beta.cell.21}}{a numeric vector}
    \item{\code{alpha.cell.108}}{a numeric vector}
    \item{\code{beta.cell.22}}{a numeric vector}
    \item{\code{alpha.cell.109}}{a numeric vector}
    \item{\code{alpha.cell.110}}{a numeric vector}
    \item{\code{alpha.cell.111}}{a numeric vector}
    \item{\code{alpha.cell.112}}{a numeric vector}
    \item{\code{gamma.cell.28}}{a numeric vector}
    \item{\code{gamma.cell.29}}{a numeric vector}
    \item{\code{gamma.cell.30}}{a numeric vector}
    \item{\code{gamma.cell.31}}{a numeric vector}
    \item{\code{beta.cell.23}}{a numeric vector}
    \item{\code{gamma.cell.32}}{a numeric vector}
    \item{\code{alpha.cell.113}}{a numeric vector}
    \item{\code{alpha.cell.114}}{a numeric vector}
    \item{\code{gamma.cell.33}}{a numeric vector}
    \item{\code{gamma.cell.34}}{a numeric vector}
    \item{\code{alpha.cell.115}}{a numeric vector}
    \item{\code{alpha.cell.116}}{a numeric vector}
    \item{\code{gamma.cell.35}}{a numeric vector}
    \item{\code{PSC.cell}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.8}}{a numeric vector}
    \item{\code{gamma.cell.36}}{a numeric vector}
    \item{\code{beta.cell.24}}{a numeric vector}
    \item{\code{beta.cell.25}}{a numeric vector}
    \item{\code{delta.cell.9}}{a numeric vector}
    \item{\code{gamma.cell.37}}{a numeric vector}
    \item{\code{beta.cell.26}}{a numeric vector}
    \item{\code{beta.cell.27}}{a numeric vector}
    \item{\code{alpha.cell.117}}{a numeric vector}
    \item{\code{alpha.cell.118}}{a numeric vector}
    \item{\code{alpha.cell.119}}{a numeric vector}
    \item{\code{alpha.cell.120}}{a numeric vector}
    \item{\code{alpha.cell.121}}{a numeric vector}
    \item{\code{beta.cell.28}}{a numeric vector}
    \item{\code{alpha.cell.122}}{a numeric vector}
    \item{\code{delta.cell.10}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.9}}{a numeric vector}
    \item{\code{acinar.cell.10}}{a numeric vector}
    \item{\code{ductal.cell.7}}{a numeric vector}
    \item{\code{gamma.cell.38}}{a numeric vector}
    \item{\code{beta.cell.29}}{a numeric vector}
    \item{\code{gamma.cell.39}}{a numeric vector}
    \item{\code{beta.cell.30}}{a numeric vector}
    \item{\code{ductal.cell.8}}{a numeric vector}
    \item{\code{alpha.cell.123}}{a numeric vector}
    \item{\code{delta.cell.11}}{a numeric vector}
    \item{\code{beta.cell.31}}{a numeric vector}
    \item{\code{alpha.cell.124}}{a numeric vector}
    \item{\code{alpha.cell.125}}{a numeric vector}
    \item{\code{alpha.cell.126}}{a numeric vector}
    \item{\code{gamma.cell.40}}{a numeric vector}
    \item{\code{alpha.cell.127}}{a numeric vector}
    \item{\code{alpha.cell.128}}{a numeric vector}
    \item{\code{alpha.cell.129}}{a numeric vector}
    \item{\code{gamma.cell.41}}{a numeric vector}
    \item{\code{alpha.cell.130}}{a numeric vector}
    \item{\code{delta.cell.12}}{a numeric vector}
    \item{\code{alpha.cell.131}}{a numeric vector}
    \item{\code{alpha.cell.132}}{a numeric vector}
    \item{\code{beta.cell.32}}{a numeric vector}
    \item{\code{alpha.cell.133}}{a numeric vector}
    \item{\code{gamma.cell.42}}{a numeric vector}
    \item{\code{alpha.cell.134}}{a numeric vector}
    \item{\code{alpha.cell.135}}{a numeric vector}
    \item{\code{alpha.cell.136}}{a numeric vector}
    \item{\code{beta.cell.33}}{a numeric vector}
    \item{\code{alpha.cell.137}}{a numeric vector}
    \item{\code{ductal.cell.9}}{a numeric vector}
    \item{\code{ductal.cell.10}}{a numeric vector}
    \item{\code{PSC.cell.1}}{a numeric vector}
    \item{\code{alpha.cell.138}}{a numeric vector}
    \item{\code{alpha.cell.139}}{a numeric vector}
    \item{\code{acinar.cell.11}}{a numeric vector}
    \item{\code{alpha.cell.140}}{a numeric vector}
    \item{\code{gamma.cell.43}}{a numeric vector}
    \item{\code{alpha.cell.141}}{a numeric vector}
    \item{\code{alpha.cell.142}}{a numeric vector}
    \item{\code{gamma.cell.44}}{a numeric vector}
    \item{\code{acinar.cell.12}}{a numeric vector}
    \item{\code{acinar.cell.13}}{a numeric vector}
    \item{\code{gamma.cell.45}}{a numeric vector}
    \item{\code{alpha.cell.143}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.10}}{a numeric vector}
    \item{\code{acinar.cell.14}}{a numeric vector}
    \item{\code{delta.cell.13}}{a numeric vector}
    \item{\code{gamma.cell.46}}{a numeric vector}
    \item{\code{alpha.cell.144}}{a numeric vector}
    \item{\code{alpha.cell.145}}{a numeric vector}
    \item{\code{delta.cell.14}}{a numeric vector}
    \item{\code{alpha.cell.146}}{a numeric vector}
    \item{\code{alpha.cell.147}}{a numeric vector}
    \item{\code{alpha.cell.148}}{a numeric vector}
    \item{\code{alpha.cell.149}}{a numeric vector}
    \item{\code{alpha.cell.150}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.11}}{a numeric vector}
    \item{\code{alpha.cell.151}}{a numeric vector}
    \item{\code{gamma.cell.47}}{a numeric vector}
    \item{\code{delta.cell.15}}{a numeric vector}
    \item{\code{gamma.cell.48}}{a numeric vector}
    \item{\code{alpha.cell.152}}{a numeric vector}
    \item{\code{acinar.cell.15}}{a numeric vector}
    \item{\code{alpha.cell.153}}{a numeric vector}
    \item{\code{ductal.cell.11}}{a numeric vector}
    \item{\code{gamma.cell.49}}{a numeric vector}
    \item{\code{alpha.cell.154}}{a numeric vector}
    \item{\code{alpha.cell.155}}{a numeric vector}
    \item{\code{mast.cell}}{a numeric vector}
    \item{\code{alpha.cell.156}}{a numeric vector}
    \item{\code{gamma.cell.50}}{a numeric vector}
    \item{\code{alpha.cell.157}}{a numeric vector}
    \item{\code{alpha.cell.158}}{a numeric vector}
    \item{\code{alpha.cell.159}}{a numeric vector}
    \item{\code{alpha.cell.160}}{a numeric vector}
    \item{\code{alpha.cell.161}}{a numeric vector}
    \item{\code{alpha.cell.162}}{a numeric vector}
    \item{\code{alpha.cell.163}}{a numeric vector}
    \item{\code{gamma.cell.51}}{a numeric vector}
    \item{\code{alpha.cell.164}}{a numeric vector}
    \item{\code{ductal.cell.12}}{a numeric vector}
    \item{\code{delta.cell.16}}{a numeric vector}
    \item{\code{beta.cell.34}}{a numeric vector}
    \item{\code{alpha.cell.165}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.12}}{a numeric vector}
    \item{\code{mast.cell.1}}{a numeric vector}
    \item{\code{beta.cell.35}}{a numeric vector}
    \item{\code{alpha.cell.166}}{a numeric vector}
    \item{\code{beta.cell.36}}{a numeric vector}
    \item{\code{alpha.cell.167}}{a numeric vector}
    \item{\code{alpha.cell.168}}{a numeric vector}
    \item{\code{mast.cell.2}}{a numeric vector}
    \item{\code{ductal.cell.13}}{a numeric vector}
    \item{\code{PSC.cell.2}}{a numeric vector}
    \item{\code{alpha.cell.169}}{a numeric vector}
    \item{\code{alpha.cell.170}}{a numeric vector}
    \item{\code{alpha.cell.171}}{a numeric vector}
    \item{\code{ductal.cell.14}}{a numeric vector}
    \item{\code{ductal.cell.15}}{a numeric vector}
    \item{\code{ductal.cell.16}}{a numeric vector}
    \item{\code{alpha.cell.172}}{a numeric vector}
    \item{\code{ductal.cell.17}}{a numeric vector}
    \item{\code{ductal.cell.18}}{a numeric vector}
    \item{\code{alpha.cell.173}}{a numeric vector}
    \item{\code{alpha.cell.174}}{a numeric vector}
    \item{\code{alpha.cell.175}}{a numeric vector}
    \item{\code{beta.cell.37}}{a numeric vector}
    \item{\code{ductal.cell.19}}{a numeric vector}
    \item{\code{alpha.cell.176}}{a numeric vector}
    \item{\code{alpha.cell.177}}{a numeric vector}
    \item{\code{alpha.cell.178}}{a numeric vector}
    \item{\code{alpha.cell.179}}{a numeric vector}
    \item{\code{delta.cell.17}}{a numeric vector}
    \item{\code{ductal.cell.20}}{a numeric vector}
    \item{\code{alpha.cell.180}}{a numeric vector}
    \item{\code{alpha.cell.181}}{a numeric vector}
    \item{\code{alpha.cell.182}}{a numeric vector}
    \item{\code{PSC.cell.3}}{a numeric vector}
    \item{\code{alpha.cell.183}}{a numeric vector}
    \item{\code{alpha.cell.184}}{a numeric vector}
    \item{\code{ductal.cell.21}}{a numeric vector}
    \item{\code{beta.cell.38}}{a numeric vector}
    \item{\code{ductal.cell.22}}{a numeric vector}
    \item{\code{alpha.cell.185}}{a numeric vector}
    \item{\code{alpha.cell.186}}{a numeric vector}
    \item{\code{alpha.cell.187}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.13}}{a numeric vector}
    \item{\code{alpha.cell.188}}{a numeric vector}
    \item{\code{alpha.cell.189}}{a numeric vector}
    \item{\code{alpha.cell.190}}{a numeric vector}
    \item{\code{acinar.cell.16}}{a numeric vector}
    \item{\code{ductal.cell.23}}{a numeric vector}
    \item{\code{ductal.cell.24}}{a numeric vector}
    \item{\code{alpha.cell.191}}{a numeric vector}
    \item{\code{beta.cell.39}}{a numeric vector}
    \item{\code{acinar.cell.17}}{a numeric vector}
    \item{\code{alpha.cell.192}}{a numeric vector}
    \item{\code{alpha.cell.193}}{a numeric vector}
    \item{\code{ductal.cell.25}}{a numeric vector}
    \item{\code{ductal.cell.26}}{a numeric vector}
    \item{\code{ductal.cell.27}}{a numeric vector}
    \item{\code{PSC.cell.4}}{a numeric vector}
    \item{\code{gamma.cell.52}}{a numeric vector}
    \item{\code{alpha.cell.194}}{a numeric vector}
    \item{\code{co.expression.cell.1}}{a numeric vector}
    \item{\code{acinar.cell.18}}{a numeric vector}
    \item{\code{ductal.cell.28}}{a numeric vector}
    \item{\code{alpha.cell.195}}{a numeric vector}
    \item{\code{PSC.cell.5}}{a numeric vector}
    \item{\code{beta.cell.40}}{a numeric vector}
    \item{\code{alpha.cell.196}}{a numeric vector}
    \item{\code{alpha.cell.197}}{a numeric vector}
    \item{\code{beta.cell.41}}{a numeric vector}
    \item{\code{ductal.cell.29}}{a numeric vector}
    \item{\code{alpha.cell.198}}{a numeric vector}
    \item{\code{alpha.cell.199}}{a numeric vector}
    \item{\code{alpha.cell.200}}{a numeric vector}
    \item{\code{ductal.cell.30}}{a numeric vector}
    \item{\code{ductal.cell.31}}{a numeric vector}
    \item{\code{ductal.cell.32}}{a numeric vector}
    \item{\code{ductal.cell.33}}{a numeric vector}
    \item{\code{alpha.cell.201}}{a numeric vector}
    \item{\code{ductal.cell.34}}{a numeric vector}
    \item{\code{acinar.cell.19}}{a numeric vector}
    \item{\code{MHC.class.II.cell}}{a numeric vector}
    \item{\code{alpha.cell.202}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.14}}{a numeric vector}
    \item{\code{alpha.cell.203}}{a numeric vector}
    \item{\code{alpha.cell.204}}{a numeric vector}
    \item{\code{alpha.cell.205}}{a numeric vector}
    \item{\code{alpha.cell.206}}{a numeric vector}
    \item{\code{alpha.cell.207}}{a numeric vector}
    \item{\code{alpha.cell.208}}{a numeric vector}
    \item{\code{alpha.cell.209}}{a numeric vector}
    \item{\code{alpha.cell.210}}{a numeric vector}
    \item{\code{ductal.cell.35}}{a numeric vector}
    \item{\code{alpha.cell.211}}{a numeric vector}
    \item{\code{alpha.cell.212}}{a numeric vector}
    \item{\code{PSC.cell.6}}{a numeric vector}
    \item{\code{acinar.cell.20}}{a numeric vector}
    \item{\code{MHC.class.II.cell.1}}{a numeric vector}
    \item{\code{alpha.cell.213}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.15}}{a numeric vector}
    \item{\code{alpha.cell.214}}{a numeric vector}
    \item{\code{ductal.cell.36}}{a numeric vector}
    \item{\code{alpha.cell.215}}{a numeric vector}
    \item{\code{ductal.cell.37}}{a numeric vector}
    \item{\code{alpha.cell.216}}{a numeric vector}
    \item{\code{ductal.cell.38}}{a numeric vector}
    \item{\code{alpha.cell.217}}{a numeric vector}
    \item{\code{delta.cell.18}}{a numeric vector}
    \item{\code{acinar.cell.21}}{a numeric vector}
    \item{\code{co.expression.cell.2}}{a numeric vector}
    \item{\code{ductal.cell.39}}{a numeric vector}
    \item{\code{acinar.cell.22}}{a numeric vector}
    \item{\code{alpha.cell.218}}{a numeric vector}
    \item{\code{alpha.cell.219}}{a numeric vector}
    \item{\code{gamma.cell.53}}{a numeric vector}
    \item{\code{co.expression.cell.3}}{a numeric vector}
    \item{\code{alpha.cell.220}}{a numeric vector}
    \item{\code{alpha.cell.221}}{a numeric vector}
    \item{\code{ductal.cell.40}}{a numeric vector}
    \item{\code{endothelial.cell.1}}{a numeric vector}
    \item{\code{alpha.cell.222}}{a numeric vector}
    \item{\code{ductal.cell.41}}{a numeric vector}
    \item{\code{alpha.cell.223}}{a numeric vector}
    \item{\code{ductal.cell.42}}{a numeric vector}
    \item{\code{acinar.cell.23}}{a numeric vector}
    \item{\code{alpha.cell.224}}{a numeric vector}
    \item{\code{ductal.cell.43}}{a numeric vector}
    \item{\code{acinar.cell.24}}{a numeric vector}
    \item{\code{beta.cell.42}}{a numeric vector}
    \item{\code{ductal.cell.44}}{a numeric vector}
    \item{\code{ductal.cell.45}}{a numeric vector}
    \item{\code{ductal.cell.46}}{a numeric vector}
    \item{\code{ductal.cell.47}}{a numeric vector}
    \item{\code{mast.cell.3}}{a numeric vector}
    \item{\code{alpha.cell.225}}{a numeric vector}
    \item{\code{ductal.cell.48}}{a numeric vector}
    \item{\code{acinar.cell.25}}{a numeric vector}
    \item{\code{gamma.cell.54}}{a numeric vector}
    \item{\code{PSC.cell.7}}{a numeric vector}
    \item{\code{alpha.cell.226}}{a numeric vector}
    \item{\code{alpha.cell.227}}{a numeric vector}
    \item{\code{ductal.cell.49}}{a numeric vector}
    \item{\code{beta.cell.43}}{a numeric vector}
    \item{\code{alpha.cell.228}}{a numeric vector}
    \item{\code{alpha.cell.229}}{a numeric vector}
    \item{\code{alpha.cell.230}}{a numeric vector}
    \item{\code{alpha.cell.231}}{a numeric vector}
    \item{\code{ductal.cell.50}}{a numeric vector}
    \item{\code{alpha.cell.232}}{a numeric vector}
    \item{\code{co.expression.cell.4}}{a numeric vector}
    \item{\code{alpha.cell.233}}{a numeric vector}
    \item{\code{gamma.cell.55}}{a numeric vector}
    \item{\code{ductal.cell.51}}{a numeric vector}
    \item{\code{alpha.cell.234}}{a numeric vector}
    \item{\code{gamma.cell.56}}{a numeric vector}
    \item{\code{beta.cell.44}}{a numeric vector}
    \item{\code{acinar.cell.26}}{a numeric vector}
    \item{\code{ductal.cell.52}}{a numeric vector}
    \item{\code{alpha.cell.235}}{a numeric vector}
    \item{\code{alpha.cell.236}}{a numeric vector}
    \item{\code{alpha.cell.237}}{a numeric vector}
    \item{\code{alpha.cell.238}}{a numeric vector}
    \item{\code{acinar.cell.27}}{a numeric vector}
    \item{\code{alpha.cell.239}}{a numeric vector}
    \item{\code{beta.cell.45}}{a numeric vector}
    \item{\code{acinar.cell.28}}{a numeric vector}
    \item{\code{ductal.cell.53}}{a numeric vector}
    \item{\code{alpha.cell.240}}{a numeric vector}
    \item{\code{delta.cell.19}}{a numeric vector}
    \item{\code{alpha.cell.241}}{a numeric vector}
    \item{\code{ductal.cell.54}}{a numeric vector}
    \item{\code{beta.cell.46}}{a numeric vector}
    \item{\code{alpha.cell.242}}{a numeric vector}
    \item{\code{acinar.cell.29}}{a numeric vector}
    \item{\code{alpha.cell.243}}{a numeric vector}
    \item{\code{alpha.cell.244}}{a numeric vector}
    \item{\code{acinar.cell.30}}{a numeric vector}
    \item{\code{ductal.cell.55}}{a numeric vector}
    \item{\code{ductal.cell.56}}{a numeric vector}
    \item{\code{alpha.cell.245}}{a numeric vector}
    \item{\code{acinar.cell.31}}{a numeric vector}
    \item{\code{alpha.cell.246}}{a numeric vector}
    \item{\code{ductal.cell.57}}{a numeric vector}
    \item{\code{ductal.cell.58}}{a numeric vector}
    \item{\code{alpha.cell.247}}{a numeric vector}
    \item{\code{acinar.cell.32}}{a numeric vector}
    \item{\code{alpha.cell.248}}{a numeric vector}
    \item{\code{alpha.cell.249}}{a numeric vector}
    \item{\code{acinar.cell.33}}{a numeric vector}
    \item{\code{acinar.cell.34}}{a numeric vector}
    \item{\code{alpha.cell.250}}{a numeric vector}
    \item{\code{PSC.cell.8}}{a numeric vector}
    \item{\code{acinar.cell.35}}{a numeric vector}
    \item{\code{alpha.cell.251}}{a numeric vector}
    \item{\code{alpha.cell.252}}{a numeric vector}
    \item{\code{ductal.cell.59}}{a numeric vector}
    \item{\code{PSC.cell.9}}{a numeric vector}
    \item{\code{ductal.cell.60}}{a numeric vector}
    \item{\code{ductal.cell.61}}{a numeric vector}
    \item{\code{alpha.cell.253}}{a numeric vector}
    \item{\code{ductal.cell.62}}{a numeric vector}
    \item{\code{ductal.cell.63}}{a numeric vector}
    \item{\code{ductal.cell.64}}{a numeric vector}
    \item{\code{acinar.cell.36}}{a numeric vector}
    \item{\code{delta.cell.20}}{a numeric vector}
    \item{\code{alpha.cell.254}}{a numeric vector}
    \item{\code{ductal.cell.65}}{a numeric vector}
    \item{\code{alpha.cell.255}}{a numeric vector}
    \item{\code{alpha.cell.256}}{a numeric vector}
    \item{\code{alpha.cell.257}}{a numeric vector}
    \item{\code{acinar.cell.37}}{a numeric vector}
    \item{\code{gamma.cell.57}}{a numeric vector}
    \item{\code{delta.cell.21}}{a numeric vector}
    \item{\code{ductal.cell.66}}{a numeric vector}
    \item{\code{ductal.cell.67}}{a numeric vector}
    \item{\code{ductal.cell.68}}{a numeric vector}
    \item{\code{alpha.cell.258}}{a numeric vector}
    \item{\code{ductal.cell.69}}{a numeric vector}
    \item{\code{ductal.cell.70}}{a numeric vector}
    \item{\code{alpha.cell.259}}{a numeric vector}
    \item{\code{ductal.cell.71}}{a numeric vector}
    \item{\code{ductal.cell.72}}{a numeric vector}
    \item{\code{co.expression.cell.5}}{a numeric vector}
    \item{\code{alpha.cell.260}}{a numeric vector}
    \item{\code{alpha.cell.261}}{a numeric vector}
    \item{\code{alpha.cell.262}}{a numeric vector}
    \item{\code{alpha.cell.263}}{a numeric vector}
    \item{\code{ductal.cell.73}}{a numeric vector}
    \item{\code{alpha.cell.264}}{a numeric vector}
    \item{\code{ductal.cell.74}}{a numeric vector}
    \item{\code{PSC.cell.10}}{a numeric vector}
    \item{\code{beta.cell.47}}{a numeric vector}
    \item{\code{alpha.cell.265}}{a numeric vector}
    \item{\code{alpha.cell.266}}{a numeric vector}
    \item{\code{alpha.cell.267}}{a numeric vector}
    \item{\code{ductal.cell.75}}{a numeric vector}
    \item{\code{alpha.cell.268}}{a numeric vector}
    \item{\code{alpha.cell.269}}{a numeric vector}
    \item{\code{alpha.cell.270}}{a numeric vector}
    \item{\code{acinar.cell.38}}{a numeric vector}
    \item{\code{ductal.cell.76}}{a numeric vector}
    \item{\code{acinar.cell.39}}{a numeric vector}
    \item{\code{alpha.cell.271}}{a numeric vector}
    \item{\code{PSC.cell.11}}{a numeric vector}
    \item{\code{PSC.cell.12}}{a numeric vector}
    \item{\code{alpha.cell.272}}{a numeric vector}
    \item{\code{alpha.cell.273}}{a numeric vector}
    \item{\code{beta.cell.48}}{a numeric vector}
    \item{\code{alpha.cell.274}}{a numeric vector}
    \item{\code{acinar.cell.40}}{a numeric vector}
    \item{\code{alpha.cell.275}}{a numeric vector}
    \item{\code{acinar.cell.41}}{a numeric vector}
    \item{\code{alpha.cell.276}}{a numeric vector}
    \item{\code{acinar.cell.42}}{a numeric vector}
    \item{\code{alpha.cell.277}}{a numeric vector}
    \item{\code{alpha.cell.278}}{a numeric vector}
    \item{\code{ductal.cell.77}}{a numeric vector}
    \item{\code{alpha.cell.279}}{a numeric vector}
    \item{\code{PSC.cell.13}}{a numeric vector}
    \item{\code{acinar.cell.43}}{a numeric vector}
    \item{\code{ductal.cell.78}}{a numeric vector}
    \item{\code{alpha.cell.280}}{a numeric vector}
    \item{\code{alpha.cell.281}}{a numeric vector}
    \item{\code{delta.cell.22}}{a numeric vector}
    \item{\code{ductal.cell.79}}{a numeric vector}
    \item{\code{alpha.cell.282}}{a numeric vector}
    \item{\code{alpha.cell.283}}{a numeric vector}
    \item{\code{ductal.cell.80}}{a numeric vector}
    \item{\code{alpha.cell.284}}{a numeric vector}
    \item{\code{ductal.cell.81}}{a numeric vector}
    \item{\code{ductal.cell.82}}{a numeric vector}
    \item{\code{beta.cell.49}}{a numeric vector}
    \item{\code{ductal.cell.83}}{a numeric vector}
    \item{\code{ductal.cell.84}}{a numeric vector}
    \item{\code{alpha.cell.285}}{a numeric vector}
    \item{\code{beta.cell.50}}{a numeric vector}
    \item{\code{gamma.cell.58}}{a numeric vector}
    \item{\code{alpha.cell.286}}{a numeric vector}
    \item{\code{ductal.cell.85}}{a numeric vector}
    \item{\code{ductal.cell.86}}{a numeric vector}
    \item{\code{ductal.cell.87}}{a numeric vector}
    \item{\code{ductal.cell.88}}{a numeric vector}
    \item{\code{gamma.cell.59}}{a numeric vector}
    \item{\code{alpha.cell.287}}{a numeric vector}
    \item{\code{beta.cell.51}}{a numeric vector}
    \item{\code{gamma.cell.60}}{a numeric vector}
    \item{\code{ductal.cell.89}}{a numeric vector}
    \item{\code{acinar.cell.44}}{a numeric vector}
    \item{\code{acinar.cell.45}}{a numeric vector}
    \item{\code{alpha.cell.288}}{a numeric vector}
    \item{\code{ductal.cell.90}}{a numeric vector}
    \item{\code{acinar.cell.46}}{a numeric vector}
    \item{\code{acinar.cell.47}}{a numeric vector}
    \item{\code{gamma.cell.61}}{a numeric vector}
    \item{\code{ductal.cell.91}}{a numeric vector}
    \item{\code{ductal.cell.92}}{a numeric vector}
    \item{\code{acinar.cell.48}}{a numeric vector}
    \item{\code{alpha.cell.289}}{a numeric vector}
    \item{\code{acinar.cell.49}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.16}}{a numeric vector}
    \item{\code{acinar.cell.50}}{a numeric vector}
    \item{\code{acinar.cell.51}}{a numeric vector}
    \item{\code{acinar.cell.52}}{a numeric vector}
    \item{\code{acinar.cell.53}}{a numeric vector}
    \item{\code{ductal.cell.93}}{a numeric vector}
    \item{\code{ductal.cell.94}}{a numeric vector}
    \item{\code{acinar.cell.54}}{a numeric vector}
    \item{\code{ductal.cell.95}}{a numeric vector}
    \item{\code{beta.cell.52}}{a numeric vector}
    \item{\code{beta.cell.53}}{a numeric vector}
    \item{\code{acinar.cell.55}}{a numeric vector}
    \item{\code{ductal.cell.96}}{a numeric vector}
    \item{\code{acinar.cell.56}}{a numeric vector}
    \item{\code{ductal.cell.97}}{a numeric vector}
    \item{\code{acinar.cell.57}}{a numeric vector}
    \item{\code{acinar.cell.58}}{a numeric vector}
    \item{\code{ductal.cell.98}}{a numeric vector}
    \item{\code{gamma.cell.62}}{a numeric vector}
    \item{\code{acinar.cell.59}}{a numeric vector}
    \item{\code{alpha.cell.290}}{a numeric vector}
    \item{\code{acinar.cell.60}}{a numeric vector}
    \item{\code{ductal.cell.99}}{a numeric vector}
    \item{\code{acinar.cell.61}}{a numeric vector}
    \item{\code{ductal.cell.100}}{a numeric vector}
    \item{\code{acinar.cell.62}}{a numeric vector}
    \item{\code{acinar.cell.63}}{a numeric vector}
    \item{\code{ductal.cell.101}}{a numeric vector}
    \item{\code{acinar.cell.64}}{a numeric vector}
    \item{\code{gamma.cell.63}}{a numeric vector}
    \item{\code{beta.cell.54}}{a numeric vector}
    \item{\code{acinar.cell.65}}{a numeric vector}
    \item{\code{acinar.cell.66}}{a numeric vector}
    \item{\code{ductal.cell.102}}{a numeric vector}
    \item{\code{acinar.cell.67}}{a numeric vector}
    \item{\code{beta.cell.55}}{a numeric vector}
    \item{\code{beta.cell.56}}{a numeric vector}
    \item{\code{acinar.cell.68}}{a numeric vector}
    \item{\code{ductal.cell.103}}{a numeric vector}
    \item{\code{acinar.cell.69}}{a numeric vector}
    \item{\code{acinar.cell.70}}{a numeric vector}
    \item{\code{ductal.cell.104}}{a numeric vector}
    \item{\code{ductal.cell.105}}{a numeric vector}
    \item{\code{acinar.cell.71}}{a numeric vector}
    \item{\code{acinar.cell.72}}{a numeric vector}
    \item{\code{acinar.cell.73}}{a numeric vector}
    \item{\code{acinar.cell.74}}{a numeric vector}
    \item{\code{alpha.cell.291}}{a numeric vector}
    \item{\code{beta.cell.57}}{a numeric vector}
    \item{\code{acinar.cell.75}}{a numeric vector}
    \item{\code{acinar.cell.76}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.17}}{a numeric vector}
    \item{\code{ductal.cell.106}}{a numeric vector}
    \item{\code{acinar.cell.77}}{a numeric vector}
    \item{\code{ductal.cell.107}}{a numeric vector}
    \item{\code{ductal.cell.108}}{a numeric vector}
    \item{\code{ductal.cell.109}}{a numeric vector}
    \item{\code{alpha.cell.292}}{a numeric vector}
    \item{\code{ductal.cell.110}}{a numeric vector}
    \item{\code{ductal.cell.111}}{a numeric vector}
    \item{\code{acinar.cell.78}}{a numeric vector}
    \item{\code{ductal.cell.112}}{a numeric vector}
    \item{\code{gamma.cell.64}}{a numeric vector}
    \item{\code{gamma.cell.65}}{a numeric vector}
    \item{\code{ductal.cell.113}}{a numeric vector}
    \item{\code{co.expression.cell.6}}{a numeric vector}
    \item{\code{ductal.cell.114}}{a numeric vector}
    \item{\code{acinar.cell.79}}{a numeric vector}
    \item{\code{ductal.cell.115}}{a numeric vector}
    \item{\code{acinar.cell.80}}{a numeric vector}
    \item{\code{ductal.cell.116}}{a numeric vector}
    \item{\code{ductal.cell.117}}{a numeric vector}
    \item{\code{beta.cell.58}}{a numeric vector}
    \item{\code{acinar.cell.81}}{a numeric vector}
    \item{\code{acinar.cell.82}}{a numeric vector}
    \item{\code{ductal.cell.118}}{a numeric vector}
    \item{\code{alpha.cell.293}}{a numeric vector}
    \item{\code{ductal.cell.119}}{a numeric vector}
    \item{\code{acinar.cell.83}}{a numeric vector}
    \item{\code{alpha.cell.294}}{a numeric vector}
    \item{\code{acinar.cell.84}}{a numeric vector}
    \item{\code{gamma.cell.66}}{a numeric vector}
    \item{\code{ductal.cell.120}}{a numeric vector}
    \item{\code{acinar.cell.85}}{a numeric vector}
    \item{\code{ductal.cell.121}}{a numeric vector}
    \item{\code{acinar.cell.86}}{a numeric vector}
    \item{\code{gamma.cell.67}}{a numeric vector}
    \item{\code{ductal.cell.122}}{a numeric vector}
    \item{\code{alpha.cell.295}}{a numeric vector}
    \item{\code{ductal.cell.123}}{a numeric vector}
    \item{\code{ductal.cell.124}}{a numeric vector}
    \item{\code{beta.cell.59}}{a numeric vector}
    \item{\code{beta.cell.60}}{a numeric vector}
    \item{\code{ductal.cell.125}}{a numeric vector}
    \item{\code{beta.cell.61}}{a numeric vector}
    \item{\code{ductal.cell.126}}{a numeric vector}
    \item{\code{alpha.cell.296}}{a numeric vector}
    \item{\code{acinar.cell.87}}{a numeric vector}
    \item{\code{ductal.cell.127}}{a numeric vector}
    \item{\code{beta.cell.62}}{a numeric vector}
    \item{\code{acinar.cell.88}}{a numeric vector}
    \item{\code{acinar.cell.89}}{a numeric vector}
    \item{\code{beta.cell.63}}{a numeric vector}
    \item{\code{acinar.cell.90}}{a numeric vector}
    \item{\code{ductal.cell.128}}{a numeric vector}
    \item{\code{acinar.cell.91}}{a numeric vector}
    \item{\code{alpha.cell.297}}{a numeric vector}
    \item{\code{gamma.cell.68}}{a numeric vector}
    \item{\code{ductal.cell.129}}{a numeric vector}
    \item{\code{alpha.cell.298}}{a numeric vector}
    \item{\code{beta.cell.64}}{a numeric vector}
    \item{\code{acinar.cell.92}}{a numeric vector}
    \item{\code{ductal.cell.130}}{a numeric vector}
    \item{\code{alpha.cell.299}}{a numeric vector}
    \item{\code{co.expression.cell.7}}{a numeric vector}
    \item{\code{ductal.cell.131}}{a numeric vector}
    \item{\code{acinar.cell.93}}{a numeric vector}
    \item{\code{beta.cell.65}}{a numeric vector}
    \item{\code{acinar.cell.94}}{a numeric vector}
    \item{\code{ductal.cell.132}}{a numeric vector}
    \item{\code{ductal.cell.133}}{a numeric vector}
    \item{\code{gamma.cell.69}}{a numeric vector}
    \item{\code{ductal.cell.134}}{a numeric vector}
    \item{\code{alpha.cell.300}}{a numeric vector}
    \item{\code{gamma.cell.70}}{a numeric vector}
    \item{\code{alpha.cell.301}}{a numeric vector}
    \item{\code{acinar.cell.95}}{a numeric vector}
    \item{\code{beta.cell.66}}{a numeric vector}
    \item{\code{ductal.cell.135}}{a numeric vector}
    \item{\code{beta.cell.67}}{a numeric vector}
    \item{\code{acinar.cell.96}}{a numeric vector}
    \item{\code{alpha.cell.302}}{a numeric vector}
    \item{\code{beta.cell.68}}{a numeric vector}
    \item{\code{beta.cell.69}}{a numeric vector}
    \item{\code{acinar.cell.97}}{a numeric vector}
    \item{\code{acinar.cell.98}}{a numeric vector}
    \item{\code{delta.cell.23}}{a numeric vector}
    \item{\code{ductal.cell.136}}{a numeric vector}
    \item{\code{beta.cell.70}}{a numeric vector}
    \item{\code{acinar.cell.99}}{a numeric vector}
    \item{\code{ductal.cell.137}}{a numeric vector}
    \item{\code{ductal.cell.138}}{a numeric vector}
    \item{\code{co.expression.cell.8}}{a numeric vector}
    \item{\code{acinar.cell.100}}{a numeric vector}
    \item{\code{ductal.cell.139}}{a numeric vector}
    \item{\code{ductal.cell.140}}{a numeric vector}
    \item{\code{acinar.cell.101}}{a numeric vector}
    \item{\code{beta.cell.71}}{a numeric vector}
    \item{\code{ductal.cell.141}}{a numeric vector}
    \item{\code{alpha.cell.303}}{a numeric vector}
    \item{\code{alpha.cell.304}}{a numeric vector}
    \item{\code{beta.cell.72}}{a numeric vector}
    \item{\code{acinar.cell.102}}{a numeric vector}
    \item{\code{alpha.cell.305}}{a numeric vector}
    \item{\code{beta.cell.73}}{a numeric vector}
    \item{\code{delta.cell.24}}{a numeric vector}
    \item{\code{PSC.cell.14}}{a numeric vector}
    \item{\code{ductal.cell.142}}{a numeric vector}
    \item{\code{acinar.cell.103}}{a numeric vector}
    \item{\code{acinar.cell.104}}{a numeric vector}
    \item{\code{ductal.cell.143}}{a numeric vector}
    \item{\code{acinar.cell.105}}{a numeric vector}
    \item{\code{ductal.cell.144}}{a numeric vector}
    \item{\code{acinar.cell.106}}{a numeric vector}
    \item{\code{beta.cell.74}}{a numeric vector}
    \item{\code{acinar.cell.107}}{a numeric vector}
    \item{\code{beta.cell.75}}{a numeric vector}
    \item{\code{acinar.cell.108}}{a numeric vector}
    \item{\code{ductal.cell.145}}{a numeric vector}
    \item{\code{ductal.cell.146}}{a numeric vector}
    \item{\code{beta.cell.76}}{a numeric vector}
    \item{\code{alpha.cell.306}}{a numeric vector}
    \item{\code{gamma.cell.71}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.18}}{a numeric vector}
    \item{\code{alpha.cell.307}}{a numeric vector}
    \item{\code{acinar.cell.109}}{a numeric vector}
    \item{\code{acinar.cell.110}}{a numeric vector}
    \item{\code{acinar.cell.111}}{a numeric vector}
    \item{\code{alpha.cell.308}}{a numeric vector}
    \item{\code{acinar.cell.112}}{a numeric vector}
    \item{\code{beta.cell.77}}{a numeric vector}
    \item{\code{ductal.cell.147}}{a numeric vector}
    \item{\code{alpha.cell.309}}{a numeric vector}
    \item{\code{co.expression.cell.9}}{a numeric vector}
    \item{\code{acinar.cell.113}}{a numeric vector}
    \item{\code{acinar.cell.114}}{a numeric vector}
    \item{\code{epsilon.cell.1}}{a numeric vector}
    \item{\code{gamma.cell.72}}{a numeric vector}
    \item{\code{acinar.cell.115}}{a numeric vector}
    \item{\code{ductal.cell.148}}{a numeric vector}
    \item{\code{ductal.cell.149}}{a numeric vector}
    \item{\code{acinar.cell.116}}{a numeric vector}
    \item{\code{beta.cell.78}}{a numeric vector}
    \item{\code{alpha.cell.310}}{a numeric vector}
    \item{\code{acinar.cell.117}}{a numeric vector}
    \item{\code{alpha.cell.311}}{a numeric vector}
    \item{\code{ductal.cell.150}}{a numeric vector}
    \item{\code{acinar.cell.118}}{a numeric vector}
    \item{\code{alpha.cell.312}}{a numeric vector}
    \item{\code{acinar.cell.119}}{a numeric vector}
    \item{\code{beta.cell.79}}{a numeric vector}
    \item{\code{acinar.cell.120}}{a numeric vector}
    \item{\code{ductal.cell.151}}{a numeric vector}
    \item{\code{ductal.cell.152}}{a numeric vector}
    \item{\code{acinar.cell.121}}{a numeric vector}
    \item{\code{co.expression.cell.10}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.19}}{a numeric vector}
    \item{\code{acinar.cell.122}}{a numeric vector}
    \item{\code{ductal.cell.153}}{a numeric vector}
    \item{\code{beta.cell.80}}{a numeric vector}
    \item{\code{beta.cell.81}}{a numeric vector}
    \item{\code{ductal.cell.154}}{a numeric vector}
    \item{\code{PSC.cell.15}}{a numeric vector}
    \item{\code{acinar.cell.123}}{a numeric vector}
    \item{\code{beta.cell.82}}{a numeric vector}
    \item{\code{gamma.cell.73}}{a numeric vector}
    \item{\code{ductal.cell.155}}{a numeric vector}
    \item{\code{alpha.cell.313}}{a numeric vector}
    \item{\code{PSC.cell.16}}{a numeric vector}
    \item{\code{alpha.cell.314}}{a numeric vector}
    \item{\code{alpha.cell.315}}{a numeric vector}
    \item{\code{alpha.cell.316}}{a numeric vector}
    \item{\code{alpha.cell.317}}{a numeric vector}
    \item{\code{alpha.cell.318}}{a numeric vector}
    \item{\code{alpha.cell.319}}{a numeric vector}
    \item{\code{alpha.cell.320}}{a numeric vector}
    \item{\code{alpha.cell.321}}{a numeric vector}
    \item{\code{alpha.cell.322}}{a numeric vector}
    \item{\code{alpha.cell.323}}{a numeric vector}
    \item{\code{alpha.cell.324}}{a numeric vector}
    \item{\code{delta.cell.25}}{a numeric vector}
    \item{\code{ductal.cell.156}}{a numeric vector}
    \item{\code{beta.cell.83}}{a numeric vector}
    \item{\code{beta.cell.84}}{a numeric vector}
    \item{\code{alpha.cell.325}}{a numeric vector}
    \item{\code{alpha.cell.326}}{a numeric vector}
    \item{\code{beta.cell.85}}{a numeric vector}
    \item{\code{ductal.cell.157}}{a numeric vector}
    \item{\code{beta.cell.86}}{a numeric vector}
    \item{\code{alpha.cell.327}}{a numeric vector}
    \item{\code{co.expression.cell.11}}{a numeric vector}
    \item{\code{alpha.cell.328}}{a numeric vector}
    \item{\code{alpha.cell.329}}{a numeric vector}
    \item{\code{co.expression.cell.12}}{a numeric vector}
    \item{\code{PSC.cell.17}}{a numeric vector}
    \item{\code{alpha.cell.330}}{a numeric vector}
    \item{\code{alpha.cell.331}}{a numeric vector}
    \item{\code{alpha.cell.332}}{a numeric vector}
    \item{\code{beta.cell.87}}{a numeric vector}
    \item{\code{alpha.cell.333}}{a numeric vector}
    \item{\code{alpha.cell.334}}{a numeric vector}
    \item{\code{alpha.cell.335}}{a numeric vector}
    \item{\code{alpha.cell.336}}{a numeric vector}
    \item{\code{alpha.cell.337}}{a numeric vector}
    \item{\code{alpha.cell.338}}{a numeric vector}
    \item{\code{alpha.cell.339}}{a numeric vector}
    \item{\code{alpha.cell.340}}{a numeric vector}
    \item{\code{alpha.cell.341}}{a numeric vector}
    \item{\code{alpha.cell.342}}{a numeric vector}
    \item{\code{alpha.cell.343}}{a numeric vector}
    \item{\code{alpha.cell.344}}{a numeric vector}
    \item{\code{beta.cell.88}}{a numeric vector}
    \item{\code{alpha.cell.345}}{a numeric vector}
    \item{\code{beta.cell.89}}{a numeric vector}
    \item{\code{alpha.cell.346}}{a numeric vector}
    \item{\code{alpha.cell.347}}{a numeric vector}
    \item{\code{alpha.cell.348}}{a numeric vector}
    \item{\code{beta.cell.90}}{a numeric vector}
    \item{\code{alpha.cell.349}}{a numeric vector}
    \item{\code{beta.cell.91}}{a numeric vector}
    \item{\code{alpha.cell.350}}{a numeric vector}
    \item{\code{alpha.cell.351}}{a numeric vector}
    \item{\code{alpha.cell.352}}{a numeric vector}
    \item{\code{alpha.cell.353}}{a numeric vector}
    \item{\code{alpha.cell.354}}{a numeric vector}
    \item{\code{alpha.cell.355}}{a numeric vector}
    \item{\code{beta.cell.92}}{a numeric vector}
    \item{\code{alpha.cell.356}}{a numeric vector}
    \item{\code{beta.cell.93}}{a numeric vector}
    \item{\code{alpha.cell.357}}{a numeric vector}
    \item{\code{alpha.cell.358}}{a numeric vector}
    \item{\code{acinar.cell.124}}{a numeric vector}
    \item{\code{ductal.cell.158}}{a numeric vector}
    \item{\code{alpha.cell.359}}{a numeric vector}
    \item{\code{alpha.cell.360}}{a numeric vector}
    \item{\code{alpha.cell.361}}{a numeric vector}
    \item{\code{co.expression.cell.13}}{a numeric vector}
    \item{\code{alpha.cell.362}}{a numeric vector}
    \item{\code{alpha.cell.363}}{a numeric vector}
    \item{\code{alpha.cell.364}}{a numeric vector}
    \item{\code{alpha.cell.365}}{a numeric vector}
    \item{\code{alpha.cell.366}}{a numeric vector}
    \item{\code{ductal.cell.159}}{a numeric vector}
    \item{\code{alpha.cell.367}}{a numeric vector}
    \item{\code{alpha.cell.368}}{a numeric vector}
    \item{\code{beta.cell.94}}{a numeric vector}
    \item{\code{beta.cell.95}}{a numeric vector}
    \item{\code{co.expression.cell.14}}{a numeric vector}
    \item{\code{alpha.cell.369}}{a numeric vector}
    \item{\code{alpha.cell.370}}{a numeric vector}
    \item{\code{alpha.cell.371}}{a numeric vector}
    \item{\code{co.expression.cell.15}}{a numeric vector}
    \item{\code{alpha.cell.372}}{a numeric vector}
    \item{\code{alpha.cell.373}}{a numeric vector}
    \item{\code{alpha.cell.374}}{a numeric vector}
    \item{\code{alpha.cell.375}}{a numeric vector}
    \item{\code{alpha.cell.376}}{a numeric vector}
    \item{\code{beta.cell.96}}{a numeric vector}
    \item{\code{alpha.cell.377}}{a numeric vector}
    \item{\code{acinar.cell.125}}{a numeric vector}
    \item{\code{alpha.cell.378}}{a numeric vector}
    \item{\code{delta.cell.26}}{a numeric vector}
    \item{\code{alpha.cell.379}}{a numeric vector}
    \item{\code{alpha.cell.380}}{a numeric vector}
    \item{\code{alpha.cell.381}}{a numeric vector}
    \item{\code{alpha.cell.382}}{a numeric vector}
    \item{\code{alpha.cell.383}}{a numeric vector}
    \item{\code{alpha.cell.384}}{a numeric vector}
    \item{\code{beta.cell.97}}{a numeric vector}
    \item{\code{alpha.cell.385}}{a numeric vector}
    \item{\code{alpha.cell.386}}{a numeric vector}
    \item{\code{ductal.cell.160}}{a numeric vector}
    \item{\code{alpha.cell.387}}{a numeric vector}
    \item{\code{alpha.cell.388}}{a numeric vector}
    \item{\code{delta.cell.27}}{a numeric vector}
    \item{\code{co.expression.cell.16}}{a numeric vector}
    \item{\code{alpha.cell.389}}{a numeric vector}
    \item{\code{alpha.cell.390}}{a numeric vector}
    \item{\code{alpha.cell.391}}{a numeric vector}
    \item{\code{alpha.cell.392}}{a numeric vector}
    \item{\code{beta.cell.98}}{a numeric vector}
    \item{\code{alpha.cell.393}}{a numeric vector}
    \item{\code{alpha.cell.394}}{a numeric vector}
    \item{\code{delta.cell.28}}{a numeric vector}
    \item{\code{alpha.cell.395}}{a numeric vector}
    \item{\code{alpha.cell.396}}{a numeric vector}
    \item{\code{beta.cell.99}}{a numeric vector}
    \item{\code{alpha.cell.397}}{a numeric vector}
    \item{\code{delta.cell.29}}{a numeric vector}
    \item{\code{alpha.cell.398}}{a numeric vector}
    \item{\code{alpha.cell.399}}{a numeric vector}
    \item{\code{alpha.cell.400}}{a numeric vector}
    \item{\code{PSC.cell.18}}{a numeric vector}
    \item{\code{ductal.cell.161}}{a numeric vector}
    \item{\code{alpha.cell.401}}{a numeric vector}
    \item{\code{alpha.cell.402}}{a numeric vector}
    \item{\code{alpha.cell.403}}{a numeric vector}
    \item{\code{alpha.cell.404}}{a numeric vector}
    \item{\code{beta.cell.100}}{a numeric vector}
    \item{\code{alpha.cell.405}}{a numeric vector}
    \item{\code{alpha.cell.406}}{a numeric vector}
    \item{\code{beta.cell.101}}{a numeric vector}
    \item{\code{beta.cell.102}}{a numeric vector}
    \item{\code{beta.cell.103}}{a numeric vector}
    \item{\code{alpha.cell.407}}{a numeric vector}
    \item{\code{beta.cell.104}}{a numeric vector}
    \item{\code{alpha.cell.408}}{a numeric vector}
    \item{\code{alpha.cell.409}}{a numeric vector}
    \item{\code{alpha.cell.410}}{a numeric vector}
    \item{\code{alpha.cell.411}}{a numeric vector}
    \item{\code{alpha.cell.412}}{a numeric vector}
    \item{\code{alpha.cell.413}}{a numeric vector}
    \item{\code{alpha.cell.414}}{a numeric vector}
    \item{\code{beta.cell.105}}{a numeric vector}
    \item{\code{ductal.cell.162}}{a numeric vector}
    \item{\code{alpha.cell.415}}{a numeric vector}
    \item{\code{alpha.cell.416}}{a numeric vector}
    \item{\code{beta.cell.106}}{a numeric vector}
    \item{\code{delta.cell.30}}{a numeric vector}
    \item{\code{alpha.cell.417}}{a numeric vector}
    \item{\code{alpha.cell.418}}{a numeric vector}
    \item{\code{alpha.cell.419}}{a numeric vector}
    \item{\code{beta.cell.107}}{a numeric vector}
    \item{\code{alpha.cell.420}}{a numeric vector}
    \item{\code{alpha.cell.421}}{a numeric vector}
    \item{\code{gamma.cell.74}}{a numeric vector}
    \item{\code{alpha.cell.422}}{a numeric vector}
    \item{\code{beta.cell.108}}{a numeric vector}
    \item{\code{beta.cell.109}}{a numeric vector}
    \item{\code{endothelial.cell.2}}{a numeric vector}
    \item{\code{acinar.cell.126}}{a numeric vector}
    \item{\code{alpha.cell.423}}{a numeric vector}
    \item{\code{alpha.cell.424}}{a numeric vector}
    \item{\code{alpha.cell.425}}{a numeric vector}
    \item{\code{alpha.cell.426}}{a numeric vector}
    \item{\code{ductal.cell.163}}{a numeric vector}
    \item{\code{beta.cell.110}}{a numeric vector}
    \item{\code{alpha.cell.427}}{a numeric vector}
    \item{\code{beta.cell.111}}{a numeric vector}
    \item{\code{alpha.cell.428}}{a numeric vector}
    \item{\code{PSC.cell.19}}{a numeric vector}
    \item{\code{alpha.cell.429}}{a numeric vector}
    \item{\code{alpha.cell.430}}{a numeric vector}
    \item{\code{alpha.cell.431}}{a numeric vector}
    \item{\code{alpha.cell.432}}{a numeric vector}
    \item{\code{alpha.cell.433}}{a numeric vector}
    \item{\code{alpha.cell.434}}{a numeric vector}
    \item{\code{alpha.cell.435}}{a numeric vector}
    \item{\code{alpha.cell.436}}{a numeric vector}
    \item{\code{alpha.cell.437}}{a numeric vector}
    \item{\code{alpha.cell.438}}{a numeric vector}
    \item{\code{beta.cell.112}}{a numeric vector}
    \item{\code{alpha.cell.439}}{a numeric vector}
    \item{\code{alpha.cell.440}}{a numeric vector}
    \item{\code{beta.cell.113}}{a numeric vector}
    \item{\code{alpha.cell.441}}{a numeric vector}
    \item{\code{beta.cell.114}}{a numeric vector}
    \item{\code{beta.cell.115}}{a numeric vector}
    \item{\code{alpha.cell.442}}{a numeric vector}
    \item{\code{gamma.cell.75}}{a numeric vector}
    \item{\code{alpha.cell.443}}{a numeric vector}
    \item{\code{alpha.cell.444}}{a numeric vector}
    \item{\code{alpha.cell.445}}{a numeric vector}
    \item{\code{alpha.cell.446}}{a numeric vector}
    \item{\code{delta.cell.31}}{a numeric vector}
    \item{\code{alpha.cell.447}}{a numeric vector}
    \item{\code{PSC.cell.20}}{a numeric vector}
    \item{\code{PSC.cell.21}}{a numeric vector}
    \item{\code{beta.cell.116}}{a numeric vector}
    \item{\code{alpha.cell.448}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.20}}{a numeric vector}
    \item{\code{alpha.cell.449}}{a numeric vector}
    \item{\code{beta.cell.117}}{a numeric vector}
    \item{\code{delta.cell.32}}{a numeric vector}
    \item{\code{beta.cell.118}}{a numeric vector}
    \item{\code{gamma.cell.76}}{a numeric vector}
    \item{\code{epsilon.cell.2}}{a numeric vector}
    \item{\code{PSC.cell.22}}{a numeric vector}
    \item{\code{delta.cell.33}}{a numeric vector}
    \item{\code{alpha.cell.450}}{a numeric vector}
    \item{\code{beta.cell.119}}{a numeric vector}
    \item{\code{alpha.cell.451}}{a numeric vector}
    \item{\code{delta.cell.34}}{a numeric vector}
    \item{\code{gamma.cell.77}}{a numeric vector}
    \item{\code{ductal.cell.164}}{a numeric vector}
    \item{\code{alpha.cell.452}}{a numeric vector}
    \item{\code{gamma.cell.78}}{a numeric vector}
    \item{\code{alpha.cell.453}}{a numeric vector}
    \item{\code{alpha.cell.454}}{a numeric vector}
    \item{\code{alpha.cell.455}}{a numeric vector}
    \item{\code{delta.cell.35}}{a numeric vector}
    \item{\code{co.expression.cell.17}}{a numeric vector}
    \item{\code{beta.cell.120}}{a numeric vector}
    \item{\code{beta.cell.121}}{a numeric vector}
    \item{\code{alpha.cell.456}}{a numeric vector}
    \item{\code{alpha.cell.457}}{a numeric vector}
    \item{\code{ductal.cell.165}}{a numeric vector}
    \item{\code{alpha.cell.458}}{a numeric vector}
    \item{\code{beta.cell.122}}{a numeric vector}
    \item{\code{gamma.cell.79}}{a numeric vector}
    \item{\code{ductal.cell.166}}{a numeric vector}
    \item{\code{alpha.cell.459}}{a numeric vector}
    \item{\code{ductal.cell.167}}{a numeric vector}
    \item{\code{alpha.cell.460}}{a numeric vector}
    \item{\code{gamma.cell.80}}{a numeric vector}
    \item{\code{beta.cell.123}}{a numeric vector}
    \item{\code{beta.cell.124}}{a numeric vector}
    \item{\code{alpha.cell.461}}{a numeric vector}
    \item{\code{beta.cell.125}}{a numeric vector}
    \item{\code{alpha.cell.462}}{a numeric vector}
    \item{\code{alpha.cell.463}}{a numeric vector}
    \item{\code{gamma.cell.81}}{a numeric vector}
    \item{\code{alpha.cell.464}}{a numeric vector}
    \item{\code{beta.cell.126}}{a numeric vector}
    \item{\code{beta.cell.127}}{a numeric vector}
    \item{\code{gamma.cell.82}}{a numeric vector}
    \item{\code{beta.cell.128}}{a numeric vector}
    \item{\code{beta.cell.129}}{a numeric vector}
    \item{\code{beta.cell.130}}{a numeric vector}
    \item{\code{delta.cell.36}}{a numeric vector}
    \item{\code{epsilon.cell.3}}{a numeric vector}
    \item{\code{beta.cell.131}}{a numeric vector}
    \item{\code{alpha.cell.465}}{a numeric vector}
    \item{\code{alpha.cell.466}}{a numeric vector}
    \item{\code{gamma.cell.83}}{a numeric vector}
    \item{\code{alpha.cell.467}}{a numeric vector}
    \item{\code{gamma.cell.84}}{a numeric vector}
    \item{\code{ductal.cell.168}}{a numeric vector}
    \item{\code{beta.cell.132}}{a numeric vector}
    \item{\code{alpha.cell.468}}{a numeric vector}
    \item{\code{endothelial.cell.3}}{a numeric vector}
    \item{\code{co.expression.cell.18}}{a numeric vector}
    \item{\code{gamma.cell.85}}{a numeric vector}
    \item{\code{ductal.cell.169}}{a numeric vector}
    \item{\code{beta.cell.133}}{a numeric vector}
    \item{\code{alpha.cell.469}}{a numeric vector}
    \item{\code{beta.cell.134}}{a numeric vector}
    \item{\code{alpha.cell.470}}{a numeric vector}
    \item{\code{PSC.cell.23}}{a numeric vector}
    \item{\code{beta.cell.135}}{a numeric vector}
    \item{\code{ductal.cell.170}}{a numeric vector}
    \item{\code{alpha.cell.471}}{a numeric vector}
    \item{\code{beta.cell.136}}{a numeric vector}
    \item{\code{acinar.cell.127}}{a numeric vector}
    \item{\code{alpha.cell.472}}{a numeric vector}
    \item{\code{gamma.cell.86}}{a numeric vector}
    \item{\code{alpha.cell.473}}{a numeric vector}
    \item{\code{beta.cell.137}}{a numeric vector}
    \item{\code{beta.cell.138}}{a numeric vector}
    \item{\code{alpha.cell.474}}{a numeric vector}
    \item{\code{PSC.cell.24}}{a numeric vector}
    \item{\code{alpha.cell.475}}{a numeric vector}
    \item{\code{endothelial.cell.4}}{a numeric vector}
    \item{\code{alpha.cell.476}}{a numeric vector}
    \item{\code{alpha.cell.477}}{a numeric vector}
    \item{\code{PSC.cell.25}}{a numeric vector}
    \item{\code{endothelial.cell.5}}{a numeric vector}
    \item{\code{alpha.cell.478}}{a numeric vector}
    \item{\code{endothelial.cell.6}}{a numeric vector}
    \item{\code{PSC.cell.26}}{a numeric vector}
    \item{\code{alpha.cell.479}}{a numeric vector}
    \item{\code{beta.cell.139}}{a numeric vector}
    \item{\code{gamma.cell.87}}{a numeric vector}
    \item{\code{PSC.cell.27}}{a numeric vector}
    \item{\code{beta.cell.140}}{a numeric vector}
    \item{\code{gamma.cell.88}}{a numeric vector}
    \item{\code{alpha.cell.480}}{a numeric vector}
    \item{\code{alpha.cell.481}}{a numeric vector}
    \item{\code{acinar.cell.128}}{a numeric vector}
    \item{\code{alpha.cell.482}}{a numeric vector}
    \item{\code{alpha.cell.483}}{a numeric vector}
    \item{\code{alpha.cell.484}}{a numeric vector}
    \item{\code{alpha.cell.485}}{a numeric vector}
    \item{\code{delta.cell.37}}{a numeric vector}
    \item{\code{beta.cell.141}}{a numeric vector}
    \item{\code{co.expression.cell.19}}{a numeric vector}
    \item{\code{PSC.cell.28}}{a numeric vector}
    \item{\code{PSC.cell.29}}{a numeric vector}
    \item{\code{alpha.cell.486}}{a numeric vector}
    \item{\code{beta.cell.142}}{a numeric vector}
    \item{\code{ductal.cell.171}}{a numeric vector}
    \item{\code{gamma.cell.89}}{a numeric vector}
    \item{\code{alpha.cell.487}}{a numeric vector}
    \item{\code{co.expression.cell.20}}{a numeric vector}
    \item{\code{gamma.cell.90}}{a numeric vector}
    \item{\code{gamma.cell.91}}{a numeric vector}
    \item{\code{alpha.cell.488}}{a numeric vector}
    \item{\code{alpha.cell.489}}{a numeric vector}
    \item{\code{ductal.cell.172}}{a numeric vector}
    \item{\code{epsilon.cell.4}}{a numeric vector}
    \item{\code{gamma.cell.92}}{a numeric vector}
    \item{\code{alpha.cell.490}}{a numeric vector}
    \item{\code{delta.cell.38}}{a numeric vector}
    \item{\code{beta.cell.143}}{a numeric vector}
    \item{\code{gamma.cell.93}}{a numeric vector}
    \item{\code{gamma.cell.94}}{a numeric vector}
    \item{\code{gamma.cell.95}}{a numeric vector}
    \item{\code{alpha.cell.491}}{a numeric vector}
    \item{\code{alpha.cell.492}}{a numeric vector}
    \item{\code{endothelial.cell.7}}{a numeric vector}
    \item{\code{alpha.cell.493}}{a numeric vector}
    \item{\code{alpha.cell.494}}{a numeric vector}
    \item{\code{alpha.cell.495}}{a numeric vector}
    \item{\code{alpha.cell.496}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.21}}{a numeric vector}
    \item{\code{alpha.cell.497}}{a numeric vector}
    \item{\code{endothelial.cell.8}}{a numeric vector}
    \item{\code{gamma.cell.96}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.22}}{a numeric vector}
    \item{\code{gamma.cell.97}}{a numeric vector}
    \item{\code{beta.cell.144}}{a numeric vector}
    \item{\code{alpha.cell.498}}{a numeric vector}
    \item{\code{alpha.cell.499}}{a numeric vector}
    \item{\code{alpha.cell.500}}{a numeric vector}
    \item{\code{delta.cell.39}}{a numeric vector}
    \item{\code{ductal.cell.173}}{a numeric vector}
    \item{\code{alpha.cell.501}}{a numeric vector}
    \item{\code{PSC.cell.30}}{a numeric vector}
    \item{\code{alpha.cell.502}}{a numeric vector}
    \item{\code{alpha.cell.503}}{a numeric vector}
    \item{\code{alpha.cell.504}}{a numeric vector}
    \item{\code{alpha.cell.505}}{a numeric vector}
    \item{\code{ductal.cell.174}}{a numeric vector}
    \item{\code{co.expression.cell.21}}{a numeric vector}
    \item{\code{alpha.cell.506}}{a numeric vector}
    \item{\code{beta.cell.145}}{a numeric vector}
    \item{\code{gamma.cell.98}}{a numeric vector}
    \item{\code{alpha.cell.507}}{a numeric vector}
    \item{\code{alpha.cell.508}}{a numeric vector}
    \item{\code{alpha.cell.509}}{a numeric vector}
    \item{\code{alpha.cell.510}}{a numeric vector}
    \item{\code{alpha.cell.511}}{a numeric vector}
    \item{\code{gamma.cell.99}}{a numeric vector}
    \item{\code{alpha.cell.512}}{a numeric vector}
    \item{\code{alpha.cell.513}}{a numeric vector}
    \item{\code{alpha.cell.514}}{a numeric vector}
    \item{\code{ductal.cell.175}}{a numeric vector}
    \item{\code{alpha.cell.515}}{a numeric vector}
    \item{\code{alpha.cell.516}}{a numeric vector}
    \item{\code{alpha.cell.517}}{a numeric vector}
    \item{\code{ductal.cell.176}}{a numeric vector}
    \item{\code{delta.cell.40}}{a numeric vector}
    \item{\code{alpha.cell.518}}{a numeric vector}
    \item{\code{alpha.cell.519}}{a numeric vector}
    \item{\code{alpha.cell.520}}{a numeric vector}
    \item{\code{gamma.cell.100}}{a numeric vector}
    \item{\code{beta.cell.146}}{a numeric vector}
    \item{\code{beta.cell.147}}{a numeric vector}
    \item{\code{gamma.cell.101}}{a numeric vector}
    \item{\code{alpha.cell.521}}{a numeric vector}
    \item{\code{alpha.cell.522}}{a numeric vector}
    \item{\code{alpha.cell.523}}{a numeric vector}
    \item{\code{co.expression.cell.22}}{a numeric vector}
    \item{\code{alpha.cell.524}}{a numeric vector}
    \item{\code{alpha.cell.525}}{a numeric vector}
    \item{\code{alpha.cell.526}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.23}}{a numeric vector}
    \item{\code{beta.cell.148}}{a numeric vector}
    \item{\code{alpha.cell.527}}{a numeric vector}
    \item{\code{beta.cell.149}}{a numeric vector}
    \item{\code{delta.cell.41}}{a numeric vector}
    \item{\code{beta.cell.150}}{a numeric vector}
    \item{\code{alpha.cell.528}}{a numeric vector}
    \item{\code{ductal.cell.177}}{a numeric vector}
    \item{\code{alpha.cell.529}}{a numeric vector}
    \item{\code{alpha.cell.530}}{a numeric vector}
    \item{\code{acinar.cell.129}}{a numeric vector}
    \item{\code{alpha.cell.531}}{a numeric vector}
    \item{\code{alpha.cell.532}}{a numeric vector}
    \item{\code{beta.cell.151}}{a numeric vector}
    \item{\code{alpha.cell.533}}{a numeric vector}
    \item{\code{alpha.cell.534}}{a numeric vector}
    \item{\code{delta.cell.42}}{a numeric vector}
    \item{\code{alpha.cell.535}}{a numeric vector}
    \item{\code{gamma.cell.102}}{a numeric vector}
    \item{\code{alpha.cell.536}}{a numeric vector}
    \item{\code{alpha.cell.537}}{a numeric vector}
    \item{\code{gamma.cell.103}}{a numeric vector}
    \item{\code{delta.cell.43}}{a numeric vector}
    \item{\code{endothelial.cell.9}}{a numeric vector}
    \item{\code{gamma.cell.104}}{a numeric vector}
    \item{\code{endothelial.cell.10}}{a numeric vector}
    \item{\code{alpha.cell.538}}{a numeric vector}
    \item{\code{alpha.cell.539}}{a numeric vector}
    \item{\code{unclassified.endocrine.cell.24}}{a numeric vector}
    \item{\code{PSC.cell.31}}{a numeric vector}
    \item{\code{gamma.cell.105}}{a numeric vector}
    \item{\code{gamma.cell.106}}{a numeric vector}
    \item{\code{alpha.cell.540}}{a numeric vector}
    \item{\code{ductal.cell.178}}{a numeric vector}
  }
}
\details{
                acinar.cell                  alpha.cell                   beta.cell
                        130                         541                         152
         co.expression.cell                  delta.cell                 ductal.cell
                         23                          44                         179
           endothelial.cell                epsilon.cell                  gamma.cell
                         11                           5                         107
                  mast.cell           MHC.class.II.cell                    PSC.cell
                          4                           2                          32
unclassified.endocrine.cell
                         25 }
\source{
E-MTAB-5061 - Single-cell RNA-seq analysis of human pancreas from healthy individuals and type 2 diabetes patients

https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-5061/

Segerstolpe, ., Palasantza, A., Eliasson, P., Andersson, E.M., Andrasson, A.C., Sun, X., Picelli, S., Sabirsh, A., Clausen, M., Bjursell, M.K. and Smith, D.M., 2016. Single-cell transcriptome profiling of human pancreatic islets in health and type 2 diabetes. Cell metabolism, 24(4), pp.593-607.
}
\examples{
cellCounts <- enumerateCellTypes(ADAPTSdata3::normalData.5061)
}
\keyword{datasets}
